% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-splat.R
\name{addDropout}
\alias{addDropout}
\title{Simulate dropout}
\usage{
addDropout(sim, params)
}
\arguments{
\item{sim}{SCESet to add dropout to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added dropout and observed counts.
}
\description{
A logistic function is used to form a relationshop between the expression
level of a gene and the probability of dropout, giving a probability for each
gene in each cell. These probabilities are used in a Bernoulli distribution
to decide which counts should be dropped.
}

