% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-functions.R
\name{addFeatureStats}
\alias{addFeatureStats}
\title{Add feature statistics}
\usage{
addFeatureStats(sce, value = c("counts", "cpm", "tpm", "fpkm"), log = FALSE,
  offset = 1, no.zeros = FALSE)
}
\arguments{
\item{sce}{SCESet to add feature statistics to.}

\item{value}{the expression value to calculate statistics for. Options are
"counts", "cpm", "tpm" or "fpkm". The values need to exist in the
given SCESet.}

\item{log}{logical. Whether to take log2 before calculating statistics.}

\item{offset}{offset to add to avoid taking log of zero.}

\item{no.zeros}{logical. Whether to remove all zeros from each feature before
calculating statistics.}
}
\value{
SCESet with additional feature statistics
}
\description{
Add additional feature statistics to an SCESet object
}
\details{
Currently adds the following statistics: mean, variance, coefficient of
variation, median and median absolute deviation. Statistics are added to
the \code{fData} slot and are named \code{stat_[log]_value_[no0]} where
\code{log} and \code{no0} are added if those arguments are true.
}
