% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-splat.R
\name{addLibSizes}
\alias{addLibSizes}
\title{Simulate library sizes}
\usage{
addLibSizes(sim, params)
}
\arguments{
\item{sim}{SCESet to add library size to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added library sizes.
}
\description{
Simulate expected library sizes from a log-normal distribution
}

