% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-splat.R
\name{addSingleCellMeans}
\alias{addSingleCellMeans}
\title{Simulate single population cell means}
\usage{
addSingleCellMeans(sim, params)
}
\arguments{
\item{sim}{SCESet to add cell means to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added cell means.
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell.
}

