% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-simulate.R
\name{bridge}
\alias{bridge}
\title{Brownian bridge}
\usage{
bridge(x = 0, y = 0, N = 5, n = 100, sigma.fac = 0.8)
}
\arguments{
\item{x}{starting value.}

\item{y}{end value.}

\item{N}{number of steps in random walk.}

\item{n}{number of points in smoothed bridge.}

\item{sigma.fac}{multiplier specifying how extreme each step can be.}
}
\value{
Vector of length n following a path from x to y.
}
\description{
Calculate a smoothed Brownian bridge between two points. A Brownian bridge is
a random walk with fixed end points.
}
