% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bringItemsForward}
\alias{bringItemsForward}
\title{Bring items forward}
\usage{
bringItemsForward(ll, items)
}
\arguments{
\item{ll}{list to adjust item order.}

\item{items}{vector of items to bring to the front. Any not in the list will
be ignored.}
}
\value{
list with selected items first
}
\description{
Move selected items to the start of a list.
}
