% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDependencies}
\alias{checkDependencies}
\title{Check dependencies}
\usage{
checkDependencies(sim.prefix = NULL, deps = NULL)
}
\arguments{
\item{sim.prefix}{prefix for a simulation to check.}

\item{deps}{vector of dependency names.}
}
\value{
TRUE invisibly if successful
}
\description{
Check suggested dependencies and prompt the user to install them if not
available
}
