% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{checkParams}
\alias{checkParams}
\title{Check splatParams object}
\usage{
checkParams(params)
}
\arguments{
\item{params}{splatParams object to check}
}
\value{
Produces error if not valid otherwise nothing
}
\description{
Check that a splatParams object has valid parameter values.
}
\details{
The following checks are made:
\itemize{
  \item{Input has "splatParams" class}
  \item{Logical parameters are logical}
  \item{Numeric parameters are numeric}
  \item{Positive numeric parameters are positive}
  \item{Integer parameters are integers}
  \item{Probability parameters are in the range 0-1}
  \item{Vector parameters are the correct length}
  \item{Vector parameters do not contain NAs}
  \item{Non-vector parameters are single values}
}
}
\examples{
checkParams(defaultParams())
}

