% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{co.var}
\alias{co.var}
\title{Calculate coefficient of variation}
\usage{
co.var(x)
}
\arguments{
\item{x}{vector of values.}
}
\value{
Value of coefficient of variation for vector
}
\description{
Implementation of the coefficient of variation
}
