% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{defaultLunParams}
\alias{defaultLunParams}
\title{Get default Lun simulation parameters}
\usage{
defaultLunParams()
}
\value{
A splatParams object containing default Lun parameters.
}
\description{
Get a splatParams object with a set of default parameters for the Lun
simulation.
}
\examples{
params <- defaultLunParams()
params
}

