% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{defaultParams}
\alias{defaultParams}
\title{Get default simulation parameters}
\usage{
defaultParams()
}
\value{
A splatParams object containing default parameters
}
\description{
Get a splatParams object with a set of default parameters that will produce a
resonable simulation of single-cell RNA-seq count data.
}
\examples{
params <- defaultParams()
params
}

