% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.assign.eqtl.effects}
\alias{eqtl.assign.eqtl.effects}
\title{Assign eGenes-eSNPs pairs and effect sizes.}
\usage{
eqtl.assign.eqtl.effects(key, snps, eQTLparams)
}
\arguments{
\item{key}{Partial eQTL key dataframe.}

\item{snps}{Dataframe of genotype information output from `vcf.parsed`.}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details. Default=`neweQTLParams()`.}
}
\value{
The key updated with assigned eQTL effects.
}
\description{
Randomly pairs N genes (eGene) a SNP (eSNP) within the window size 
(eqtl.dist) and assigns each pair an effect size sampled from a gamma 
distribution parameterized using the effect sizes from a bulk eQTL study 
using the GTEx data from the thyroid tissue.
}
