% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.key}
\alias{eqtl.key}
\title{Select eGenes-eSNPs pairs and assign effect sizes.}
\usage{
eqtl.key(genes, snps, eQTLparams)
}
\arguments{
\item{genes}{Dataframe with gene ID and location}

\item{snps}{Dataframe with SNP ID, location, and sample genotypes}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details.}
}
\value{
A dataframe eSNPs-eGenes pair assignments and their effect sizes
}
\description{
Randomly pairs N eSNPs with an eGene within the designated window size 
(eqtl.dist) and assigns each pair an effect size sampled from a gamma 
distribution parameterized using the effect sizes from a bulk eQTL study 
using the GTEx data from the thyroid tissue.
}
