% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.parse.gff}
\alias{eqtl.parse.gff}
\title{Process gene data}
\usage{
eqtl.parse.gff(gff)
}
\arguments{
\item{gff}{Dataframe containing the genes to include in GFF/GTF format.}
}
\value{
A dataframe containing gene IDs, chr, and location.
}
\description{
Select features labeled "gene", pull the chromosome and transcriptional start
site (considering strand direction) for each gene.
}
