% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.parse.vcf}
\alias{eqtl.parse.vcf}
\title{Format and subset genotype data from a VCF file.}
\usage{
eqtl.parse.vcf(vcf, eQTLparams)
}
\arguments{
\item{vcf}{Data.frame of genotypes for samples to simulate from a VCF file.}

\item{eQTLparams}{eQTLparams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details. Default=`neweQTLParams()`.}
}
\value{
A list containing the genotype dataframe and a list of samples.
}
\description{
Convert [0/0, 0/1, 1/1] alleles into [0,1,2] format and remove SNPs that do
not meet Minor Allele Frequency requirements specified in `eQTLParams`.
}
