% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.sim.eqtl.eff}
\alias{eqtl.sim.eqtl.eff}
\title{Add eQTL effects to means matrix}
\usage{
eqtl.sim.eqtl.eff(id, key, snps, MeansPop)
}
\arguments{
\item{id}{The group ID (e.g. "global" or "g1")}

\item{key}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}

\item{snps}{A dataframe containing SNP names, locations, & sample genotypes.}

\item{MeansPop}{Mean gene expression dataframe}
}
\value{
Matrix of simulated gene means for eQTL population.
}
\description{
Add eQTL effects to means matrix
}
\details{
For eSNP-eGene pairs for id group, the eQTL effect is added to the gene mean
y = (scaled Effect Size) \* (assigned mean for that gene) \* genotype + error
}
