% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.sim.means}
\alias{eqtl.sim.means}
\title{Simulate mean gene expression matrix without eQTL effects}
\usage{
eqtl.sim.means(samples, key)
}
\arguments{
\item{samples}{Vector with the names of the samples}

\item{key}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}
}
\value{
Mean gene expression levels for the population WITHOUT eQTL effects
}
\description{
Gene mean expression levels are assigned to each gene:sample pair randomly
from a normal distribution parameterized using the mean and CV assigned to 
each gene (see `eqtl.sample.means`).
}
