% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-estimate.R
\name{eQTLEstES}
\alias{eQTLEstES}
\title{Estimate eQTL Effect Size parameters}
\usage{
eQTLEstES(all.pairs, eQTLparams)
}
\arguments{
\item{all.pairs}{Txt file with all eQTL pairs from a real eQTL analysis. Must
include 3 columns: gene_id, pval_nominal, and slope.}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details.}
}
\value{
eQTLparams object with estimated values.
}
\description{
Estimate rate and shape parameters for the gamma distribution used to
simulate eQTL (eSNP-eGene) effect sizes.
}
\details{
Parameters for the gamma distribution are estimated by fitting the top eSNP-
eGene pair effect sizes using \code{\link[fitdistrplus]{fitdist}}. The 
maximum goodness-of-fit estimation method is used to minimise the 
Cramer-von Mises distance. This can fail in some situations, in which case 
the method of moments estimation method is used instead.
}
