% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-estimate.R
\name{eQTLEstimate}
\alias{eQTLEstimate}
\title{Estimate eQTL simulation parameters}
\usage{
eQTLEstimate(
  eQTLparams = neweQTLParams(),
  all.pairs = ex_pairs,
  gene.means = ex_means
)
}
\arguments{
\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details.}

\item{all.pairs}{Txt file with all eQTL pairs from a real eQTL analysis. Must
include 3 columns: 'gene_id', 'pval_nominal', and 'slope'.}

\item{gene.means}{Dataframe of real gene means across a population, where 
each row is a gene and each column is an individual in the population.}
}
\value{
eQTLParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the eQTL population simulation from
real data. See the individual estimation functions for more details on
how this is done.
}
\examples{
# Load example data
data(ex_means)
data(ex_pairs)
eQTLparams <- eQTLEstimate()
eQTLparams

}
\seealso{
\code{\link{eQTLEstES}},  \code{\link{eQTLEstMeanCV}}
}
