% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLGeneMeans}
\alias{eQTLGeneMeans}
\title{Set a gene mean expression value (not normalized) for each gene.}
\usage{
eQTLGeneMeans(params, pairs, eQTLparams)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{pairs}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details.}
}
\value{
the eSNP-eGene pairs dataframe updated to include the mean gene 
expression level, BCV, and standard deviation.
}
\description{
Assign a mean expression value to each gene, sampled from a gamma 
distribution parameterized by splatEstimate, then calculate the BCV and
the standard deviation
}
