% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLMeansMatrix}
\alias{eQTLMeansMatrix}
\title{Project normalized gene expression matrix into mean gene expression matrix}
\usage{
eQTLMeansMatrix(pairs, nGeneMeansPop, params)
}
\arguments{
\item{pairs}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}

\item{nGeneMeansPop}{The normalized gene expression means for the population}

\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}
}
\value{
MeansMatrix: Matrix of simulated gene means for eQTL population.
}
\description{
Project normalized gene expression matrix into mean gene expression matrix
}
\details{
For each gene/sample, the normalized expression value (from rnorm) is 
transformed to the cumulative density function (pnorm) between 0 and 1, this
value is then quantile normalized (qgamma) using the gamma distribution 
parameterized from splatEstimate().
}
