% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLnormMeansMatrix}
\alias{eQTLnormMeansMatrix}
\title{Generate normalized mean gene expression matrix for whole eQTL population}
\usage{
eQTLnormMeansMatrix(id, snps, key, error, last_norm_matrix)
}
\arguments{
\item{id}{The group ID (e.g. "g1")}

\item{snps}{The dataframe with the genetic marker info}

\item{key}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}

\item{last_norm_matrix}{Dataframe of normalized means if they were already
simulated for another group.}
}
\value{
normGeneMeansPop: matrix of normalized mean gene exp. levels.
}
\description{
Use the approach outlined in Huang et. al 2018 (NAR) to assign normalized
mean expression levels for each gene for each sample. Where:
y = Effect Size * genotype + error, where error ~ Norm(0,1)
}
