% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{eQTLParams}
\alias{eQTLParams}
\alias{eQTLParams-class}
\title{The eQTLParams class}
\description{
S4 class that holds parameters for the eQTL simulation.
}
\section{Parameters}{


The eQTL simulation requires the following parameters:

\describe{
    \item{\code{[eqtl.n]}}{The number genes to assign eQTL effects.}
    \item{\code{[eqtl.dist]}}{Maximum distance between eSNP and eGene}
    \item{\code{[eqtl.maf.min]}}{Minimum Minor Allele Frequency of eSNPs.}
    \item{\code{[eqtl.maf.max]}}{Maximum Minor Allele Frequency of eSNPs.}
    \item{\code{[eqtl.groups]}}{Number of groups to simulate.}
    \item{\code{[eqtl.group.specific]}}{Percent of eQTL effects to simulate 
    as group specific.}
    \item{\emph{eQTL Effect size distribution parameters. Defaults estimated
    from GTEx eQTL mapping results, see vignette for more information.}}{
        \describe{
            \item{\code{eqtlES.shape}}{Shape parameter for the effect size 
            gamma distribution.}
            \item{\code{eqtlES.rate}}{Rate parameter for the effect size 
            gamma distribution.}
        }
    }
    \item{\emph{Bulk Mean Expression distribution parameters. Defaults 
    estimated from GTEx data, see vignette for more information.}}{
        \describe{
            \item{\code{pop.mean.shape}}{Shape parameter for the mean (i.e. 
            bulk) expression gamma distribution}
            \item{\code{pop.mean.rate}}{Rate parameter for the mean (i.e. 
            bulk) expression gamma distribution}
        }
    }
    \item{\emph{Bulk Expression Coefficient of Variation distribution
    parameters binned. Defaults estimated from GTEx data, see vignette for 
    more information.}}{
        \describe{
            \item{\code{pop.cv.param}}{Dataframe containing gene
            mean bin range, and the CV shape, and CV rate parameters for
            each of those bins.}
        }
    }
}
The parameters not shown in brackets can be estimated from real data using
\code{\link{eQTLEstimate}}. For details of the eQTL simulation
see \code{\link{eQTLSimulate}}.
}

