% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLSimulate}
\alias{eQTLSimulate}
\title{eQTL-Simulate}
\usage{
eQTLSimulate(
  params = newSplatParams(),
  eQTLparams = neweQTLParams(),
  gff = ex_gff,
  vcf = ex_snps,
  key = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details. Default=`newSplatParams()`.}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details. Default=`neweQTLParams()`.}

\item{gff}{Data.frame of genes to simulate from a GFF or GTF file.}

\item{vcf}{Data.frame of genotypes for samples to simulate from a VCF file.}

\item{key}{Data.frame of complete or partial eQTL key that is output from 
`eQTLSimulate()`. If FALSE, a key will be generated from scratch. 
Default=FALSE.}

\item{verbose}{logical. Whether to print progress messages. Default=TRUE.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
A list containing: `means` a dataframe or list of dataframes (if 
n.groups > 1) with the simulated mean gene expression value for each gene 
(row) and each sample (column) and `key` a dataframe with the eSNP-eGene
assignments.
}
\description{
Simulate mean gene counts for a population of samples based on sample 
genotype with eQTL effects included for certain genes (eGenes).
}
\details{
SplatParams and eQTLParams can be set in a variety of ways. 1. If 
not provided, default parameters are used. 2. Default parameters can be 
overridden by supplying desired parameters using \code{\link{setParams}}. 
3. Parameters can be estimated from real data of your choice using 
\code{\link{splatEstimate}} and \code{\link{eQTLEstimate}}.

`eQTLSimulate` involves the following steps:
\enumerate{
    \item Load eQTL key or generate eQTL key from the GTF/GFF file.
    \item Format and subset genotype data from the VCF file.
    \item If not in key, assign expression mean and variance to each gene.
    \item If not in key, assign eGenes-eSNPs pairs and effect sizes.
    \item If not in key and groups >1, assign subset of eQTL associations as 
    group-specific. 
    \item Simulate mean gene expression matrix without eQTL effects
    \item Add eQTL effects to means matrix.
    \item Quantile normalize by sample to fit single-cell expression 
    distribution as defined in `splatEstimate`.
    \item Add quantile normalized gene mean and cv info the eQTL key.}
}
\examples{
# Load example data
data(ex_gff)
data(ex_snps)
pop.sim <- eQTLSimulate()

}
\seealso{
\code{\link{eqtl.parse.gff}}, \code{\link{eqtl.parse.vcf}},
\code{\link{eqtl.assign.means}}, \code{\link{eqtl.assign.eqtl.effects}}, 
\code{\link{eqtl.assign.group.effects}}, \code{\link{eqtl.sim.means}},
\code{\link{eqtl.sim.eqtl.eff}}, \code{\link{eqtl.quan.norm.sc}}, 
\code{\link{eqtl.key.update.quan.norm}}
}
