% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLsnps}
\alias{eQTLsnps}
\title{Process genotype data}
\usage{
eQTLsnps(vcf, eQTLparams)
}
\arguments{
\item{vcf}{Dataframe containing real/simulated genotype data in .vcf format. 
Where each column is a sample and each row is a SNP.}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details.}
}
\value{
A dataframe containing SNP names, locations, and sample genotypes.
}
\description{
Read in SNP (genotype matrix) file and remove extra columns. Then calculate
the Minor Allele Frequency and filter SNPs outside of the MAF range defined.
}
