% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estBCVParams}
\alias{estBCVParams}
\title{Estimate Biological Coefficient of Variation parameters}
\usage{
estBCVParams(counts, params)
}
\arguments{
\item{counts}{counts matrix to estimate parameters from.}

\item{params}{splatParams object to store estimated values in.}
}
\value{
spaltParams object with estimated values.
}
\description{
Parameters are estimated using the \code{estimateDisp} function in the
\code{edgeR} package. Specifically the common dispersion and prior degrees
of freedom. See \code{\link{estimateDisp}} for details.
}
\examples{
\dontrun{
data("sc_example_counts")
params <- splatParams()
params <- estBCVParams(sc_example_counts, params)
params
}
}

