% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estDropoutParams}
\alias{estDropoutParams}
\title{Estimate dropout parameters}
\usage{
estDropoutParams(norm.counts, params)
}
\arguments{
\item{norm.counts}{library size normalised counts matrix.}

\item{params}{splatParams object to store estimated values in.}
}
\value{
Params object with estimated values.
}
\description{
Estimate the midpoint and shape parameters for the logistic function used
when simulating dropout. Also estimates whether dropout is likely to be
present in the dataset.
}
\details{
Logistic function parameters are estimated by fitting a logistic function
to the relationship between log2 mean gene expression and the proportion of
zeros in each gene. See \code{\link[stats]{nls}} for details of fitting. The
presence of dropout is determined by comparing the observed number of zeros
in each gene to the expected number of zeros from a negative binomial
distribution with the gene mean and a dispersion of 0.1. If the maximum
difference between the observed number of zeros and the expected number is
greater than 10 percent of the number of cells
(\code{max(obs.zeros - exp.zeros) > 0.1 * ncol(norm.counts)}) then dropout is
considered to be present in the dataset. This is a somewhat crude measure
but should give a reasonable indication. A more accurate approach is to look
at a plot of log2 mean expression vs the difference between observed and
expected number of zeros across all genes.
}
\examples{
\dontrun{
data("sc_example_counts")
norm_ex_counts <- t(t(sc_example_counts) / colSums(sc_example_counts) *
                  median(colSums(sc_example_counts)))
params <- splatParams()
params <- estMeanParams(norm_ex_counts, params)
}
}

