% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateParams}
\alias{estimateParams}
\alias{estimateParams.SCESet}
\alias{estimateParams.matrix}
\title{Estimate simulation parameters}
\usage{
estimateParams(x, params = NULL)

\method{estimateParams}{SCESet}(x, params = NULL)

\method{estimateParams}{matrix}(x, params = NULL)
}
\arguments{
\item{x}{either a counts matrix or an SCESet object containing count data to
estimate parameters from.}

\item{params}{splatParams object to store estimated values in. If not
provided a new object will be created.}
}
\value{
A params object containing the estimated parameters.
}
\description{
Estimate simulation parameters from a real dataset. See the individual
estimation functions for more details on how this is done.
}
\examples{
data("sc_example_counts")
params <- estimateParams(sc_example_counts)
params
# Replace defaults with estimated params
params <- defaultParams()
params <- estimateParams(sc_example_counts, params)
params
}
\seealso{
\code{\link{estMeanParams}},  \code{\link{estLibParams}},
\code{\link{estOutlierParams}}, \code{\link{estBCVParams}},
\code{\link{estDropoutParams}}
}

