% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-estimate.R
\name{estimateSplatParams}
\alias{estimateSplatParams}
\alias{estimateSplatParams.SCESet}
\alias{estimateSplatParams.matrix}
\title{Estimate Splatter simulation parameters}
\usage{
estimateSplatParams(counts, params = newSplatParams())

\method{estimateSplatParams}{SCESet}(counts, params = newSplatParams())

\method{estimateSplatParams}{matrix}(counts, params = newSplatParams())
}
\arguments{
\item{counts}{either a counts matrix or an SCESet object containing count
data to estimate parameters from.}

\item{params}{SplatParams object to store estimated values in.}
}
\value{
SplatParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the Splatter simulation from a real
dataset. See the individual estimation functions for more details on how this
is done.
}
\examples{
data("sc_example_counts")
params <- estimateSplatParams(sc_example_counts)
params
}
\seealso{
\code{\link{estSplatMean}},  \code{\link{estSplatLib}},
\code{\link{estSplatOutlier}}, \code{\link{estSplatBCV}},
\code{\link{estSplatDropout}}
}

