% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estLibParams}
\alias{estLibParams}
\title{Estimate library size parameters}
\usage{
estLibParams(counts, params)
}
\arguments{
\item{counts}{counts matrix to estimate parameters from.}

\item{params}{splatParams object to store estimated values in.}
}
\value{
splatParams object with estimated values.
}
\description{
A log-normal distribution is fitted to the library sizes and the estimated
parameters are added to the params object. See \code{\link{fitdist}} for
details on the fitting.
}
\examples{
data("sc_example_counts")
params <- params()
params <- estLibParams(sc_example_counts, params)
params
}

