% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estOutlierParams}
\alias{estOutlierParams}
\title{Estimate expression outlier parameters}
\usage{
estOutlierParams(norm.counts, params)
}
\arguments{
\item{norm.counts}{library size normalised counts matrix.}

\item{params}{splatParams object to store estimated values in.}
}
\value{
splatParams object with estimated values.
}
\description{
Parameters are estimated by comparing means of individual genes to the
median mean expression level.
}
\details{
Expression outlier genes are detected using the Median Absolute Deviation
(MAD) from median method. If the log2 mean expression of a gene is greater
than two MADs from the median log2 mean expression it is designated as a
outlier. The proportion of outlier genes is used to estimate the outlier
probability. The low outlier probability is estimated as the proportion of
outlier genes that have a log2 mean less than the median log2 mean. Factors
for each outlier gene are calculated by dividing mean expression by the
median mean expression. A log-normal distribution is then fitted to these
factors in order to estimate the outlier factor location and scale
parameters. See \code{\link[fitdistrplus]{fitdist}} for details on the
fitting.
}
\examples{
\dontrun{
data("sc_example_counts")
norm_ex_counts <- t(t(sc_example_counts) / colSums(sc_example_counts) *
                  median(colSums(sc_example_counts)))
params <- splatParams()
params <- estOutlierParams(norm_ex_counts, params)
params
}
}

