% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-estimate.R
\name{estSplatMeans}
\alias{estSplatMeans}
\title{Estimate Splatter mean parameters}
\usage{
estSplatMeans(norm.counts, params)
}
\arguments{
\item{norm.counts}{library size normalised counts matrix.}

\item{params}{SplatParams object to store estimated values in.}
}
\value{
SplatParams object with estimated values.
}
\description{
Estimate rate and shape parameters for the gamma distribution used to
simulate gene expression means using the 'moment matching estimation' method
of \code{\link[fitdistrplus]{fitdist}}.
}

