% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_gff}
\alias{ex_gff}
\title{Example GFF file}
\format{A data frame with 504 rows and 9 variables:
\describe{
  \item{V1}{id}
  \item{V2}{source}
  \item{V3}{feature}
  \item{V4}{start}
  \item{V5}{end}
  \item{V6}{score}
  \item{V7}{direction}
  \item{V8}{frame}
  \item{V9}{attribute}
  ...
}}
\source{
\url{ftp://ftp.ensembl.org/pub/release-98/gff3/homo_sapiens/Homo_sapiens.GRCh38.98.chromosome.22.gff3.gz
}
}
\usage{
data(ex_gff)
}
\description{
An example GFF file containing human genes from chromosome 22 of the GRCh38
}
\keyword{datasets}
