% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_means}
\alias{ex_means}
\title{Example gene mean data for a whole population (from GTEx)}
\format{
A data frame with 2438 rows and 850 variables:
}
\source{
\url{https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_tpm.gct.gz
}
}
\usage{
data(ex_means)
}
\description{
An example population wide gene mean data.frame, where each column is a 
sample and each row is a gene. Data is from the Thyroid tissue from the GTEx
dataset. Example data has already been filtered to remove genes with >50% 
of samples have < 0.1 TPM
}
\keyword{datasets}
