% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_pairs}
\alias{ex_pairs}
\title{Example eQTL results from GTEx Thyroid tissue}
\format{A data frame with 25,244 rows and 9 variables:}
\source{
\url{https://storage.googleapis.com/gtex_analysis_v7/single_tissue_eqtl_data/all_snp_gene_associations/Thyroid.allpairs.txt.gz
}
}
\usage{
data(ex_pairs)
}
\description{
An example dataframe with eQTL pairs and effect sizes from the Thyroid 
tissue from the GTEx dataset. This example data has already been filtered to 
include only the top eSNP for each gene. Columns that must be included are: 
'gene_id', 'pval_nominal', and 'slope'.
}
\keyword{datasets}
