% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_snps}
\alias{ex_snps}
\title{Example simulated genotype file (vcf)}
\format{
A data frame with 141882 rows and 109 variables:
}
\source{
\url{ftp://ftp.ensembl.org/pub/release-98/gff3/homo_sapiens/Homo_sapiens.GRCh38.98.chromosome.22.gff3.gz
}
}
\usage{
data(ex_snps)
}
\description{
An example genotype file in the .vcf format, where each column is a sample
and each row is a SNP.
}
\keyword{datasets}
