% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplatParams-methods.R
\docType{methods}
\name{expandParams,SplatParams-method}
\alias{expandParams,SplatParams-method}
\title{Expand parameters}
\usage{
\S4method{expandParams}{SplatParams}(object)
}
\arguments{
\item{object}{splatParams object to expand.}
}
\value{
Expanded splatParams object.
}
\description{
Expand the parameters that can be vectors so that they are the same length as
the number of groups.
}
\examples{
\dontrun{
params <- newSplatParams()
params <- setParams(params, groupCells = c(10, 10))
params
params <- expandParams(params)
params
}
}

