% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SplatParams-methods.R
\docType{methods}
\name{expandParams}
\alias{expandParams}
\alias{expandParams,SplatParams-method}
\title{Expand parameters}
\usage{
expandParams(object, ...)

\S4method{expandParams}{SplatParams}(object)
}
\arguments{
\item{object}{object to expand.}

\item{...}{additional arguments.}
}
\value{
Expanded object.
}
\description{
Expand the parameters that can be vectors so that they are the same length as
the number of groups.
}

