% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{expandPathParams}
\alias{expandPathParams}
\title{Expath path parameters}
\usage{
expandPathParams(params)
}
\arguments{
\item{params}{splatParams object to expand.}
}
\value{
expanded splatParams object.
}
\description{
Expand the path parameters so that they are the same length as the number
of groups.
}
\examples{
params <- defaultParams()
params <- setParams(params, groupCells = c(10, 10))
params
params <- expandPathParams(params)
params
}

