% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTeamRows.R
\name{generateTeamRows}
\alias{generateTeamRows}
\title{Create all select inputs for comp analyzer}
\usage{
generateTeamRows(team_data, weapon_types, session)
}
\description{
Create all select inputs for comp analyzer
}
