% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeaponRow.R
\name{generateWeaponRow}
\alias{generateWeaponRow}
\title{Function to generate shiny UI for one weapon for team comp analyzer}
\usage{
generateWeaponRow(
  weapon_types,
  session,
  item_team_id,
  item_player_id,
  item_weapon
)
}
\description{
Function to generate shiny UI for one weapon for team comp analyzer
}
