% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeaponValueRowSlider.R
\name{generateWeaponValueRowSlider}
\alias{generateWeaponValueRowSlider}
\title{Function to generate sliders for weapon values on weapon value table}
\usage{
generateWeaponValueRowSlider(val, label, min_val, max_val, session)
}
\description{
Function to generate sliders for weapon values on weapon value table
}
