% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeaponValueTable.R
\name{generateWeaponValueTable}
\alias{generateWeaponValueTable}
\title{Table to display weapon values}
\usage{
generateWeaponValueTable(df, mode_type)
}
\description{
Table to display weapon values
}
