% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCE-functions.R
\name{getCounts}
\alias{getCounts}
\title{Get counts}
\usage{
getCounts(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
Counts matrix
}
\description{
Get counts matrix from a SingleCellExperiment object. If counts is missing
a warning is issued and the first assay is returned.
}
