% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-simulate.R
\name{getExpSttFactors}
\alias{getExpSttFactors}
\title{Get exponential of Student's t-factors (i.e. log-Student's t)}
\usage{
getExpSttFactors(n.facs, sel.prob, neg.prob, fac.loc, fac.scale, fac.nu,
  fac.min, fac.max)
}
\arguments{
\item{n.facs}{Number of factors to generate.}

\item{sel.prob}{Probability that a factor will be selected to be different
from 1.}

\item{neg.prob}{Probability that a selected factor is less than one.}

\item{fac.loc}{Location parameter for the log-Stt distribution.}

\item{fac.scale}{Scale factor for the log-Stt distribution.}

\item{fac.nu}{Dispersion parameter for the log-Stt distribution.}

\item{fac.min}{Minimum fold change value}

\item{fac.max}{Maximum fold change value}
}
\value{
Vector containing generated factors.
}
\description{
Randomly generate multiplication factors from a log-Stt distribution.
}
