% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Params-methods.R
\name{getParam}
\alias{getParam}
\alias{getParam,Params-method}
\title{Get a parameter}
\usage{
getParam(object, name)

\S4method{getParam}{Params}(object, name)
}
\arguments{
\item{object}{object to get parameter from.}

\item{name}{name of the parameter to get.}
}
\value{
The extracted parameter value
}
\description{
Accessor function for getting parameter values.
}
\examples{
params <- newSimpleParams()
getParam(params, "nGenes")

}
