% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{getParams}
\alias{getParams}
\title{Get parameters from splatParams object}
\usage{
getParams(params, names)
}
\arguments{
\item{params}{splatParams object to get parameters from.}

\item{names}{vector of parameter names to extract.}
}
\value{
Vector if all selected parameters are single values, otherwise a
        list.
}
\description{
Get values for the parameters in a splatParams object. Uses the same pattern
(category.parameter) as \code{\link{setParams}}.
}
\examples{
params <- defaultParams()
# Get the number of genes
getParams(params, "nGenes")
# Get the number of genes and the mean rate parameter
getParams(params, c("nGenes", "mean.rate"))
# Returns a list if one of the selected parameters is a vector
params <- setParams(params, groupCells = c(100, 200))
getParams(parans, c("nGenes", "mean.rate", "groupCells"))
}

