% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{getPathOrder}
\alias{getPathOrder}
\title{Get path order}
\usage{
getPathOrder(path.from)
}
\arguments{
\item{path.from}{Vector giving the path endpoints that each path orginates
from.}
}
\value{
Vector giving the order to process paths in.
}
\description{
Identify the correct order to process paths so that preceding paths have
already been simulated.
}
\examples{
path.order <- getPathOrder(c(2, 0, 2))
}

