% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kersplat-estimate.R
\name{kersplatEstBCV}
\alias{kersplatEstBCV}
\title{Estimate Kersplat BCV parameters}
\usage{
kersplatEstBCV(counts, params, verbose)
}
\arguments{
\item{counts}{counts matrix.}

\item{params}{KersplatParams object to store estimated values in.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
KersplatParams object with estimated BCV parameters
}
\description{
Estimate Biological Coefficient of Variation (BCV) parameters for the
Kersplat simulation
}
\details{
The \code{\link[edgeR]{estimateDisp}} function is used to estimate the common
dispersion across the dataset. An exponential correction is applied based on
fitting an exponential relationship between simulated and estimated values.
If this results in a negative dispersion a simpler linear correction is
applied instead.
}
