% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kersplat-simulate.R
\name{kersplatSelectRegs}
\alias{kersplatSelectRegs}
\title{Select Kersplat regulators}
\usage{
kersplatSelectRegs(params, verbose)
}
\arguments{
\item{params}{KersplatParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
KersplatParams object with gene regulators
}
\description{
Select regulator genes in the gene network for a Kersplat simulation
}
\details{
Regulators are randomly selected, weighted according to the difference
between their out degree and in degree. This is an arbitrary weighting and
may be improved or replace in the future.
}
