% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kersplat-simulate.R
\name{kersplatSimAmbientCounts}
\alias{kersplatSimAmbientCounts}
\title{Simulate Kersplat ambient counts}
\usage{
kersplatSimAmbientCounts(sim, params, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment containing simulation.}

\item{params}{KersplatParams object with simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SingleCellExperiment with ambient counts
}
\description{
Simulate Kersplat ambient counts
}
\details{
The overall expression profile to calculated by averaging the cell counts
of the (non-empty) cells. This is then multiplied by the ambient library
sizes to get a mean for each cell. Counts are then sampled from a Poisson
distribution using these means.
}
