% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kersplat-simulate.R
\name{kersplatSimCellCounts}
\alias{kersplatSimCellCounts}
\title{Simulate Kersplat cell counts}
\usage{
kersplatSimCellCounts(sim, params, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment containing simulation.}

\item{params}{KersplatParams object with simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SingleCellExperiment with cell counts
}
\description{
Simulate cell counts for the Kersplat simulation
}
\details{
Counts are sampled from a Poisson distribution with lambda equal to the
cell means matrix.
}
