% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{kersplatSimCounts}
\alias{kersplatSimCounts}
\title{Simulate Kersplat final counts}
\usage{
kersplatSimCounts(sim, params, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment containing simulation.}

\item{params}{SplotParams object with simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SingleCellExperiment with counts matrix
}
\description{
Simulate the final counts matrix for a Kersplat simulation
}
\details{
The cell counts matrix and ambient counts matrix are added together. The
result is then downsampled to the cell library size (for cells and doublets)
or the ambient library size (for empty cells) using the
\code{\link[DropletUtils]{downsampleMatrix}} function.
}
\seealso{
\code{\link[DropletUtils]{downsampleMatrix}}
}
