% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSims.R
\name{listSims}
\alias{listSims}
\title{List simulations}
\usage{
listSims(print = TRUE)
}
\arguments{
\item{print}{logical. Whether to print to the console.}
}
\value{
Invisibly returns a data.frame containing the information that is
displayed.
}
\description{
List all the simulations that are currently available in Splatter with a
brief description.
}
\examples{
listSims()
}
