% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lun2-simulate.R
\name{lun2Simulate}
\alias{lun2Simulate}
\title{Lun2 simulation}
\usage{
lun2Simulate(params = newLun2Params(), zinb = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{params}{Lun2Params object containing simulation parameters.}

\item{zinb}{logical. Whether to use a zero-inflated model.}

\item{verbose}{logical. Whether to print progress messages}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SCESet containing simulated counts.
}
\description{
Simulate single-cell RNA-seq count data using the method described in Lun
and Marioni "Overcoming confounding plate effects in differential expression
analyses of single-cell RNA-seq data".
}
\details{

}
\examples{
sim <- lun2Simulate()
}

