% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{makeCompPanel}
\alias{makeCompPanel}
\title{Make comparison panel}
\usage{
makeCompPanel(comp, labels = c("A", "B", "C", "D", "E", "F", "G", ""))
}
\arguments{
\item{comp}{list returned by \code{\link{combineSCESets}}.}
}
\value{
Combined panel plot
}
\description{
Combine the plots from \code{combineSCESets} into a single panel.
}
\examples{
sim1 <- splatSimulate(nGenes = 1000, groupCells = 20)
sim2 <- simpleSimulate(nGenes = 1000, nCells = 20)
comparison <- compareSCESets(list(Splat = sim1, Simple = sim2))
panel <- makeCompPanel(comparison)

}
