% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{makeOverallPanel}
\alias{makeOverallPanel}
\title{Make overall panel}
\usage{
makeOverallPanel(comp, diff)
}
\arguments{
\item{comp}{list returned by \code{\link{compSCESets}}.}

\item{diff}{list returned by \code{\link{diffSCESets}}.}
}
\value{
Combined panel plot
}
\description{
Combine the plots from \code{compSCESets} and \code{diffSCESets} into a
single panel.
}
\examples{
sim1 <- splatSimulate(nGenes = 1000, groupCells = 20)
sim2 <- simpleSimulate(nGenes = 1000, nCells = 20)
comparison <- compSCESets(list(Splat = sim1, Simple = sim2))
difference <- diffSCESets(list(Splat = sim1, Simple = sim2), ref = "Simple")
panel <- makeCompPanel(comparison, difference)

}
