% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{mergeParams}
\alias{mergeParams}
\title{Merge two splatParams objects}
\usage{
mergeParams(params1, params2)
}
\arguments{
\item{params1}{first splatParams object to merge.}

\item{params2}{second splatParams object to merge.}
}
\value{
Merged splatParams object.
}
\description{
Merge two splatParams objects. Any parameters that are NA in the first
splatParams object are replaced by the value in the second splatParams
object.
}
\examples{
params <- splatParams(nGenes = 1000, mean.rate = 0.5)
params
# Replace unset parameters with default parameters
params <- mergeParams(params, defaultParams())
params
}

