% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mock_bulk_eqtl}
\alias{mock_bulk_eqtl}
\title{Generate mock eQTL mapping results}
\usage{
mock_bulk_eqtl(n_genes = 1000)
}
\arguments{
\item{n_genes}{Number of genes in mock eQTL data. Default=1,000.}
}
\value{
data.frame containing mock bulk eQTL mapping results.
}
\description{
Quick function to generate mock eQTL mapping results, with parameters 
estimated using real eQTL mapping results from GTEX using thyroid tissue.
}
