% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mock_bulk_matrix}
\alias{mock_bulk_matrix}
\title{Generate mock bulk population scale expression data}
\usage{
mock_bulk_matrix(n_genes = 1000, n_samples = 100)
}
\arguments{
\item{n_genes}{Number of genes in mock bulk data. Default=1,000.}

\item{n_samples}{Number of samples in mock bulk data. Default=100.}
}
\value{
data.frame containing mock bulk expression data.
}
\description{
Quick function to generate mock bulk expression data for a population, with
parameters estimated using real thyroid tissue data from GTEx.
}
