% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mock_gff}
\alias{mock_gff}
\title{Generate mock gff}
\usage{
mock_gff(n_genes = 1000, chromosome = 1)
}
\arguments{
\item{n_genes}{Number of genes in mock gff file}

\item{chromosome}{Chromosome value. Default=1.}
}
\value{
data.frame containing mock gff data from chromosome 1.
}
\description{
Quick function to generate a mock gff.
}
