% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mock_vcf}
\alias{mock_vcf}
\title{Generate mock vcf}
\usage{
mock_vcf(n_snps = 10000, n_samples = 10, chromosome = 1)
}
\arguments{
\item{n_snps}{Number of SNPs in mock vcf file. Default=10,000.}

\item{n_samples}{Number of samples in mock bulk data. Default=10.}

\item{chromosome}{Chromosome value. Default=1.}
}
\value{
data.frame containing mock gff data from chromosome 1.
}
\description{
Quick function to generate mock vcf file. Note this data has unrealistic
population structure.
}
