% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mockBulkeQTL}
\alias{mockBulkeQTL}
\title{Generate mock eQTL mapping results}
\usage{
mockBulkeQTL(n.genes = 1000)
}
\arguments{
\item{n.genes}{Number of genes in mock eQTL data.}
}
\value{
data.frame containing mock bulk eQTL mapping results.
}
\description{
Quick function to generate mock eQTL mapping results, with parameters
estimated using real eQTL mapping results from GTEx using thyroid tissue.
}
\examples{
eqtl <- mockBulkeQTL()

}
