% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mockVCF}
\alias{mockVCF}
\title{Generate mock vcf}
\usage{
mockVCF(
  n.snps = 200,
  n.samples = 5,
  chromosome = 1,
  chr.length = 2e+06,
  seed = NULL
)
}
\arguments{
\item{n.snps}{Number of SNPs in mock vcf file.}

\item{n.samples}{Number of samples in mock bulk data.}

\item{chromosome}{Chromosome name}

\item{chr.length}{Length of mock chromosome}

\item{seed}{Optional: seed for random seed}
}
\value{
data.frame containing mock gff data.
}
\description{
Quick function to generate mock vcf file. Note this data has unrealistic
population structure.
}
\examples{
vcf <- mockVCF()

}
