% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleParams-methods.R
\name{newSimpleParams}
\alias{newSimpleParams}
\title{New SimpleParams}
\usage{
newSimpleParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}
}
\value{
SimpleParams object.
}
\description{
Create a new SimpleParams object.
}
\examples{
params <- newSimpleParams()
}

