% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplatParams-methods.R
\name{newSplatParams}
\alias{newSplatParams}
\title{New SplatParams}
\usage{
newSplatParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}
}
\value{
SplatParams object.
}
\description{
Create a new SplatParams object.
}
\examples{
params <- newSplatParams()
}

