% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-simulate.R
\name{pop.assign.eqtl.effects}
\alias{pop.assign.eqtl.effects}
\title{Assign eGenes-eSNPs pairs and effect sizes.}
\usage{
pop.assign.eqtl.effects(key, snps, popParams)
}
\arguments{
\item{key}{Partial eQTL key dataframe.}

\item{snps}{Dataframe of genotype information output from `vcf.parsed`.}

\item{popParams}{popParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{popParams}} for details. Default=`newPopParams()`.}
}
\value{
The key updated with assigned eQTL effects.
}
\description{
Randomly pairs N genes (eGene) a SNP (eSNP) within the window size 
(eqtl.dist) and assigns each pair an effect size sampled from a gamma 
distribution parameterized using the effect sizes from a real eQTL study.
}
