% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-simulate.R
\name{pop.assign.group.effects}
\alias{pop.assign.group.effects}
\title{Designate subset of eQTL associations as group-specific.}
\usage{
pop.assign.group.effects(key, groups, popParams, params)
}
\arguments{
\item{key}{Partial eQTL key dataframe.}

\item{groups}{array of group names}

\item{popParams}{popParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{popParams}} for details. Default=`newPopParams()`.}

\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details. Default=`newSplatParams()`.}
}
\value{
he key updated with group eQTL and non-eQTL effects.
}
\description{
If groups > 1, n eSNP-eGene pairs (n = 'eqtl.group.specific') are randomly
assigned as group specific.
}
