% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-simulate.R
\name{pop.assign.means}
\alias{pop.assign.means}
\title{Assign expression mean and variance to each gene}
\usage{
pop.assign.means(params, key, popParams)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{key}{Partial eQTL key dataframe.}

\item{popParams}{popParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{popParams}} for details.}
}
\value{
The key updated with assigned means and variances.
}
\description{
A mean and coefficient of variation is assigned to each gene by sampling from
gamma distributions parameterized from real data using `splatEstimate`. 
The cv gamma distributions are binned by gene mean because the distribution 
of variance in real data is not independent from the mean.
}
