% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-simulate.R
\name{pop.key.update.quan.norm}
\alias{pop.key.update.quan.norm}
\title{Add quantile normalized gene mean and cv info the eQTL key.}
\usage{
pop.key.update.quan.norm(key, MeansMatrix)
}
\arguments{
\item{key}{Partial eQTL key dataframe.}

\item{MeansMatrix}{The output from `pop.quan.norm.sc` as a matrix or list of
matrices.}
}
\value{
Final eQTL key.
}
\description{
Add quantile normalized gene mean and cv info the eQTL key.
}
