% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-simulate.R
\name{pop.parse.gff}
\alias{pop.parse.gff}
\title{Generate eQTL key matrix using information from the GTF/GFF file.}
\usage{
pop.parse.gff(gff)
}
\arguments{
\item{gff}{Dataframe of genes to simulate from a GFF or GTF file.}
}
\value{
The partial eQTL key dataframe.
}
\description{
Generate eQTL key matrix using information from the GTF/GFF file.
}
