% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-simulate.R
\name{pop.parse.vcf}
\alias{pop.parse.vcf}
\title{Format and subset genotype data from a VCF file.}
\usage{
pop.parse.vcf(vcf, popParams)
}
\arguments{
\item{vcf}{Data.frame of genotypes for samples to simulate from a VCF file.}

\item{popParams}{popParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{popParams}} for details. Default=`newPopParams()`.}
}
\value{
A list containing the genotype dataframe and a list of samples.
}
\description{
Convert [0/0, 0/1, 1/1] alleles into [0,1,2] format and remove SNPs that do
not meet Minor Allele Frequency requirements specified in `popParams`.
}
