% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-simulate.R
\name{pop.quan.norm.sc}
\alias{pop.quan.norm.sc}
\title{Quantile normalize by sample to fit sc expression distribution.}
\usage{
pop.quan.norm.sc(params, MeansMatrix)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details. Default=`newSplatParams()`.}

\item{MeansMatrix}{Mean gene expression with eQTL effects dataframe.}
}
\value{
Dataframe of quantile normalized gene mean expression levels.
}
\description{
For each sample, expression values are quantile normalized (qgamma) 
using the gamma distribution parameterized from splatEstimate(). This ensures
the simulated gene means reflect the distribution expected from a sc dataset
and not a bulk dataset.
}
