% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-estimate.R
\name{popEstimate.ES}
\alias{popEstimate.ES}
\title{Estimate eQTL Effect Size parameters}
\usage{
popEstimate.ES(eqtl, popParams)
}
\arguments{
\item{eqtl}{Txt file with all or top eQTL pairs from a real eQTL analysis. 
Must include columns: gene_id, pval_nominal, and slope.}

\item{popParams}{popParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{popParams}} for details.}
}
\value{
popParams object with estimated values.
}
\description{
Estimate rate and shape parameters for the gamma distribution used to
simulate eQTL (eSNP-eGene) effect sizes.
}
\details{
Parameters for the gamma distribution are estimated by fitting the top eSNP-
eGene pair effect sizes using \code{\link[fitdistrplus]{fitdist}}. The 
maximum goodness-of-fit estimation method is used to minimise the 
Cramer-von Mises distance. This can fail in some situations, in which case 
the method of moments estimation method is used instead.
}
