% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print.splatParams}
\alias{print.splatParams}
\title{Print splatParams object}
\usage{
\method{print}{splatParams}(x, ...)
}
\arguments{
\item{x}{splatParams object to print.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns x (via \code{\link{invisible}(x)})
}
\description{
Pretty print the parameters stored in a splatParams object. Parameters are
labelled as either (estimatable) or [not estimatable].
}
\examples{
params <- defaultParams()
print(params)
}

