% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{quantileNormalizeSC}
\alias{quantileNormalizeSC}
\title{Project normalized gene expression matrix into mean gene expression matrix}
\usage{
quantileNormalizeSC(params, MeansMatrix)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{MeansMatrix}{The gene expression means for the population}
}
\value{
MeansMatrix: Matrix of simulated gene means for eQTL population.
}
\description{
Project normalized gene expression matrix into mean gene expression matrix
}
\details{
For each sample, expression value is quantile normalized (qgamma) using the
gamma distribution parameterized from splatEstimate().
}
