% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbindMatched}
\alias{rbindMatched}
\title{Bind rows (matched)}
\usage{
rbindMatched(df1, df2)
}
\arguments{
\item{df1}{first data.frame to bind.}

\item{df2}{second data.frame to bind.}
}
\value{
data.frame containing rows from \code{df1} and \code{df2} but only
        common columns.
}
\description{
Bind the rows of two data frames, keeping only the columns that are common
to both.
}
