% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kersplat-estimate.R
\name{selectFit}
\alias{selectFit}
\title{Select fit}
\usage{
selectFit(data, distr, weights = NULL, verbose = TRUE)
}
\arguments{
\item{data}{The data to fit}

\item{distr}{Name of the distribution to fit}

\item{weights}{Optional vector of weights}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
The selected fit object
}
\description{
Try a variety of fitting methods and select the best one
}
\details{
The distribution is fitted to the data using each of the
\code{\link[fitdistrplus]{fitdist}} fitting methods. The fit with the
smallest Cramer-von Mises statistic is selected.
}
