% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Lun2Params-methods.R,
%   R/LunParams-methods.R, R/Params-methods.R, R/PhenoParams-methods.R,
%   R/SCDDParams-methods.R, R/SplatParams-methods.R
\docType{methods}
\name{setParam}
\alias{setParam}
\alias{setParam,Lun2Params-method}
\alias{setParam,LunParams-method}
\alias{setParam,Params-method}
\alias{setParam,PhenoParams-method}
\alias{setParam,SCDDParams-method}
\alias{setParam,SplatParams-method}
\title{Set a parameter}
\usage{
setParam(object, name, value)

\S4method{setParam}{Lun2Params}(object, name, value)

\S4method{setParam}{LunParams}(object, name, value)

\S4method{setParam}{Params}(object, name, value)

\S4method{setParam}{PhenoParams}(object, name, value)

\S4method{setParam}{SCDDParams}(object, name, value)

\S4method{setParam}{SplatParams}(object, name, value)
}
\arguments{
\item{object}{object to set parameter in.}

\item{name}{name of the parameter to set.}

\item{value}{value to set the paramter to.}
}
\value{
Object with new parameter value.
}
\description{
Function for setting parameter values.
}
\examples{
params <- newSimpleParams()
setParam(params, "nGenes", 100)

}
