% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{setParams}
\alias{setParams}
\title{Update a splatParams object}
\usage{
setParams(params, ...)
}
\arguments{
\item{params}{the splatParams object to update.}

\item{...}{Any parameters to set.}
}
\value{
splatParms object with updated parameters
}
\description{
Set any of the parameters in a splatParams object to have a new value.
}
\details{
This function allows multiple parameters to be updated or set using a single
simple function call. Parameters to update are specified by supplying
additional arguments that follow the levels of the splatParams data structure
separated by the "." character. For example
\code{setParams(params, nGenes = 100)} is equivalent to
\code{params$nGenes <- 100} and \code{update(params, mean.rate = 1)} is
equivalent to \code{params$mean$rate <- 1}. For more details of the available
parameters and the splatParams data structure see \code{\link{splatParams}}.
}
\examples{
params <- defaultParams()
params
# Set nGenes
params <- setParams(params, nGenes = 1000)
params
# Set mean rate parameter and library size location parameter
params <- setParams(params, mean.rate = 1, lib.loc = 12)
params
}

