% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-functions.R
\name{showDFs}
\alias{showDFs}
\title{Show data.frame}
\usage{
showDFs(dfs, not.default)
}
\arguments{
\item{dfs}{list of data.frames to show.}

\item{not.default}{logical vector giving which have changed from the default.}
}
\value{
Print data.frame parameters
}
\description{
Function used for pretty printing data.frame parameters.
}
