% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-splat.R
\name{simBCVMeans}
\alias{simBCVMeans}
\title{Simulate BCV means}
\usage{
simBCVMeans(sim, params)
}
\arguments{
\item{sim}{SCESet to add BCV means to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added BCV means.
}
\description{
Simulate means for each gene in each cell that are adjusted to follow a
mean-variance trend using Biological Coefficient of Variation taken from
and inverse gamma distribution.
}

