% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-splat.R
\name{simGeneMeans}
\alias{simGeneMeans}
\title{Simulate gene means}
\usage{
simGeneMeans(sim, params)
}
\arguments{
\item{sim}{SCESet to add gene means to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added gene means.
}
\description{
Simulate gene means from a gamma distribution. Also simulates outlier
expression factors. Genes with an outlier factor not equal to 1 are replaced
with the median mean expression multiplied by the outlier factor.
}

