% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simGroupCellMeans}
\alias{simGroupCellMeans}
\title{Simulate group cell means}
\usage{
simGroupCellMeans(sim, params)
}
\arguments{
\item{sim}{SCESet to add cell means to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added cell means.
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell. Cells start with the mean expression for the group they belong to
which is adjusted for each cells expected library size.
}

