% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simGroupDE}
\alias{simGroupDE}
\title{Simulate group differential expression}
\usage{
simGroupDE(sim, params)
}
\arguments{
\item{sim}{SCESet to add differential expression to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added differential expression.
}
\description{
Simulate differential expression for groups. Differential expression
factors for each group are produced using \code{\link{getLNormFactors}} and
these are added along with updated means for each group.
}

