% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simPathCellMeans}
\alias{simPathCellMeans}
\title{Simulate path cell means}
\usage{
simPathCellMeans(sim, params)
}
\arguments{
\item{sim}{SCESet to add cell means to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added cell means.
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell. Cells are assigned assigned a random position on the appropriate
path. The mean at that position is then adjusted for each cells expected
library size.
}

