% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-splat.R
\name{simPathDE}
\alias{simPathDE}
\title{Simulate path differential expression}
\usage{
simPathDE(sim, params)
}
\arguments{
\item{sim}{SCESet to add differential expression to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SCESet with added differential expression.
}
\description{
Simulate differential expression for path. Similar to
\code{\link{simGroupDE}} but care has to be taken to make sure paths are
processed in the correct order.
}

