% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SimpleParams}
\alias{SimpleParams}
\alias{SimpleParams-class}
\title{The SimpleParams class}
\description{
S4 class that holds parameters for the simple simulation.
}
\section{Parameters}{


The simple simulation uses the following parameters:

\describe{
\item{\code{nGenes}}{The number of genes to simulate.}
\item{\code{nCells}}{The number of cells to simulate.}
\item{\code{[seed]}}{Seed to use for generating random numbers.}
\item{\code{mean.shape}}{The shape parameter for the mean gamma
distribution.}
\item{\code{mean.rate}}{The rate parameter for the mean gamma
distribution.}
\item{\code{[count.disp]}}{The dispersion parameter for the counts
negative binomial distribution.}
}

The parameters not shown in brackets can be estimated from real data using
\code{\link{simpleEstimate}}. For details of the simple simulation
see \code{\link{simpleSimulate}}.
}

