% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-simple.R
\name{simSimple}
\alias{simSimple}
\title{Simple simulation}
\usage{
simSimple(params = defaultParams(), verbose = TRUE, ...)
}
\arguments{
\item{params}{splatParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SCESet containing simulated counts
}
\description{
Simulate counts from a simple negative binomial distribution without
simulated library sizes, differential expression etc.
}
\details{
Uses the following parameters: \code{nCells}, \code{nGenes},
\code{mean.shape}, \code{mean.rate}, \code{bcv.common}.

Gene means are simulated from a gamma distribution with
\code{shape = mean.shape} and \code{rate = mean.rate}. Counts are then
simulated from a negative binomial distribution with \code{mu = means} and
\code{size = 1 / bcv.common}.

Parameters are set in the tiered manner described in \code{\link{splat}}.
}
\examples{
sim <- simSimple()
}

