% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{splat}
\alias{splat}
\title{FUNCTION TITLE}
\usage{
splat(params = defaultParams(), method = c("groups", "paths"),
  add.assay = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{params}{DESCRIPTION.}

\item{method}{DESCRIPTION.}

\item{add.assay}{DESCRIPTION.}

\item{verbose}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN DESCRIPTION
}
\description{
FUNCTION DESCRIPTION
}
\examples{
# ADD EXAMPLES HERE
}

