% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-eqtl.R
\name{splateQTL}
\alias{splateQTL}
\title{Splat-eQTL}
\usage{
splateQTL(
  params = newSplatParams(),
  gff_file = "test_data/test.gff3",
  snp_file = "test_data/test.vcf",
  eqtlES_shape = 2.740558,
  eqtlES_rate = 6.441281,
  esnp.n = 100,
  eqtl.dist = 1e+06,
  eqtl.maf = 0.1,
  eqtl.mafd = 0.01,
  eqtl.save = TRUE,
  ...
)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{gff_file}{Path to a GFF/GTF file containing genes to include.}

\item{snp_file}{Path to real/simulated genotype data in .vcf format. 
Where each column is a sample and each row is a SNP.}

\item{eqtlES_shape}{Effect Size shape parameter (default estimated from 
GTEx thyroid cis-eQTL data)}

\item{eqtlES_rate}{Effect Size rate parameter (default estimated from 
GTEx thyroid cis-eQTL data)}

\item{esnp.n}{Number of eSNP-eQTL associations to include}

\item{eqtl.dist}{Distance between eSNP and eGene (TSS).}

\item{eqtl.maf}{Desired Minor Allele Frequency (MAF) of eSNPs to include}

\item{eqtl.mafd}{Maximum variation from eqtl.maf to include as eSNP}

\item{eqtl.save}{logical. Whether to save eQTL key and mean matrix.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
GeneMeansPop Matrix containing the simulated mean gene expression
value for each gene (row) and each sample in the population (column).
intermediate values.
}
\description{
Simulate mean gene counts for a population of samples, such that a defined 
number of associations (cis-eQTL) between markers (eSNPs) and genes 
(eGenes) exist.
}
\details{
Parameters can be set in a variety of ways. If no parameters are provided
the default parameters are used. Any parameters in \code{params} can be
overridden by supplying additional arguments through a call to
\code{\link{setParams}}. This design allows the user flexibility in
how they supply parameters and allows small adjustments without creating a
new \code{SplatParams} object. See examples for a demonstration of how this
can be used.

The eQTL Gene Mean simulation involves the following steps:
\enumerate{
    \item Load and format gene (GFF/GTF) and SNP (genotype) data.
    \item Select eGenes-eSNPs pairs and assign effect sizes.
    \item Generate normalized gene mean expression matrix for the population. 
    \item Set a gene mean expression value (not normalized) for each gene. 
    \item Generate a gene mean expression matrix for the population.
    \item (optional) Save eQTL key (pairs)
}
}
\examples{
# Load example data
library(scater)
set.seed(1)
sce <- mockSCE()
params <- splatEstimate(sce)
pop.gMeans <- splateQTL(params)

}
\seealso{
\code{\link{splateQTLgenes}}, \code{\link{splateQTLsnps}},
\code{\link{splateQTLpairs}}, \code{\link{splateQTLnormMeansMatrix}},
\code{\link{splateQTLGeneMeans}}, \code{\link{splateQTLMeansMatrix}}
}
