% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-eqtl.R
\name{splateQTLGeneMeans}
\alias{splateQTLGeneMeans}
\title{Set a gene mean expression value (not normalized) for each gene.}
\usage{
splateQTLGeneMeans(params, pairs)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{pairs}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}
}
\value{
the eSNP-eGene pairs dataframe updated to include the mean gene 
expression level, BCV, and standard deviation.
}
\description{
Assign a mean expression value to each gene, sampled from a gamma 
distribution parameterized by splatEstimate, then calculate the BCV and
the standard deviation
}
