% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-eqtl.R
\name{splateQTLgenes}
\alias{splateQTLgenes}
\title{Process gene data}
\usage{
splateQTLgenes(gff)
}
\arguments{
\item{gff}{Dataframe containing the genes to include in GFF/GTF format.}
}
\value{
A dataframe containing gene IDs and locations.
}
\description{
Read in GFF/GTF file (ignoring header) and select only sequences where the 
feature is listed as a gene. Then get the Transcriptional Start Site for 
each gene (depending on strand direction).
}
