% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-eqtl.R
\name{splateQTLMeansMatrix}
\alias{splateQTLMeansMatrix}
\title{Un-normalize the mean gene expression matrix}
\usage{
splateQTLMeansMatrix(pairs, nMeans)
}
\arguments{
\item{nMeans}{The normalized gene expression means for the population}

\item{params}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}
}
\value{
MeansMatrix: Matrix of simulated gene means for eQTL population.
}
\description{
Un-normalize the mean gene expression matrix
}
\details{
For each gene/sample, the normalized expression value (from rnorm) is 
transformed to the cumulative density function (pnorm) between 0 and 1, this
value is then inversed (qnorm) to map the probability to a value defined by
the gene mean assigned in splateQTLGeneMeans.
}
