% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-eqtl.R
\name{splateQTLpairs}
\alias{splateQTLpairs}
\title{Select eGenes-eSNPs pairs and assign effect sizes.}
\usage{
splateQTLpairs(genes, snps, esnp.n, eqtl.dist, eqtlES_shape, eqtlES_rate)
}
\arguments{
\item{genes}{Dataframe with gene ID and location}

\item{snps}{Dataframe with SNP ID, location, and sample genotypes}

\item{esnp.n}{Number of eSNP-eQTL associations to include}

\item{eqtl.dist}{Distance between eSNP and eGene (TSS).}

\item{eqtlES_shape}{Effect Size shape parameter (default estimated from 
GTEx thyroid cis-eQTL data)}

\item{eqtlES_rate}{Effect Size rate parameter (default estimated from 
GTEx thyroid cis-eQTL data)}
}
\value{
A dataframe eSNPs-eGenes pair assignments and their effect sizes
}
\description{
Randomly pairs N eSNPs with an eGene within the designated window size 
(eqtl.dist) and assigns each pair an effect size sampled from a gamma 
distribution parameterized using the effect sizes from a bulk eQTL study 
using the GTEx data from the thyroid tissue.
}
