% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-eqtl.R
\name{splateQTLsnps}
\alias{splateQTLsnps}
\title{Process genotype data}
\usage{
splateQTLsnps(snps, eqtl.maf, eqtl.mafd)
}
\arguments{
\item{snps}{Dataframe containing real/simulated genotype data in .vcf format. 
Where each column is a sample and each row is a SNP.}

\item{eqtl.maf}{Desired Minor Allele Frequency (MAF) of eSNPs to include}

\item{eqtl.mafd}{Maximum variation from eqtl.maf to include as eSNP}
}
\value{
A dataframe containing SNP names, locations, and sample genotypes.
}
\description{
Read in SNP (genotype matrix) file and remove extra columns. Then calculate
the Minor Allele Frequency and filter SNPs outside of the MAF range defined.
}
