% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-estimate.R
\name{splatEstBCV}
\alias{splatEstBCV}
\title{Estimate Splat Biological Coefficient of Variation parameters}
\usage{
splatEstBCV(counts, params)
}
\arguments{
\item{counts}{counts matrix to estimate parameters from.}

\item{params}{SplatParams object to store estimated values in.}
}
\value{
SplatParams object with estimated values.
}
\description{
Parameters are estimated using the \code{estimateDisp} function in the
\code{edgeR} package. Specifically the common dispersion and prior degrees
of freedom. See \code{\link{estimateDisp}} for details.
}
