% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-estimate.R
\name{splatEstDropout}
\alias{splatEstDropout}
\title{Estimate Splat dropout parameters}
\usage{
splatEstDropout(norm.counts, params)
}
\arguments{
\item{norm.counts}{library size normalised counts matrix.}

\item{params}{SplatParams object to store estimated values in.}
}
\value{
SplatParams object with estimated values.
}
\description{
Estimate the midpoint and shape parameters for the logistic function used
when simulating dropout.
}
\details{
Logistic function parameters are estimated by fitting a logistic function
to the relationship between log2 mean gene expression and the proportion of
zeros in each gene. See \code{\link[stats]{nls}} for details of fitting.
Note this is done on the experiment level, more granular (eg. group or cell)
level dropout is not estimated.
}
