% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-estimate.R
\name{splatEstimate}
\alias{splatEstimate}
\alias{splatEstimate.SCESet}
\alias{splatEstimate.matrix}
\title{Estimate Splat simulation parameters}
\usage{
splatEstimate(counts, params = newSplatParams())

\method{splatEstimate}{SCESet}(counts, params = newSplatParams())

\method{splatEstimate}{matrix}(counts, params = newSplatParams())
}
\arguments{
\item{counts}{either a counts matrix or an SCESet object containing count
data to estimate parameters from.}

\item{params}{SplatParams object to store estimated values in.}
}
\value{
SplatParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the Splat simulation from a real
dataset. See the individual estimation functions for more details on how this
is done.
}
\examples{
data("sc_example_counts")
params <- splatEstimate(sc_example_counts)
params
}
\seealso{
\code{\link{splatEstMean}},  \code{\link{splatEstLib}},
\code{\link{splatEstOutlier}}, \code{\link{splatEstBCV}},
\code{\link{splatEstDropout}}
}

