% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopCleanSCE}
\alias{splatPopCleanSCE}
\title{Clean up the population-scale SCE to remove redundant information}
\usage{
splatPopCleanSCE(sim.all)
}
\arguments{
\item{sim.all}{SingleCellExperiment object with counts for all samples}
}
\value{
SingleCellExperiment with simulated sc counts.
}
\description{
Clean up the population-scale SCE to remove redundant information
}
