% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopConditionalEffects}
\alias{splatPopConditionalEffects}
\title{Add conditional DE effects to means matrix}
\usage{
splatPopConditionalEffects(id, key, vcf, means.pop)
}
\arguments{
\item{id}{The group ID (e.g. "global" or "g1")}

\item{key}{Partial splatPop key data.frame.}

\item{vcf}{VariantAnnotation object containing genotypes of samples.}

\item{means.pop}{Population mean gene expression matrix}
}
\value{
data.frame of gene mean expression levels WITH eQTL effects.
}
\description{
Add conditional DE effects to means matrix
}
\keyword{internal}
