% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopConditionEffects}
\alias{splatPopConditionEffects}
\title{Assign Condition-specific eQTL and DEGs.}
\usage{
splatPopConditionEffects(params, key, conditions)
}
\arguments{
\item{params}{SplatPopParams object containing parameters for population
scale simulations. See \code{\link{SplatPopParams}} for details.}

\item{key}{Partial splatPop key data.frame.}

\item{conditions}{array of condition names}
}
\value{
The key updated with conditional eQTL and DE effects.
}
\description{
If nConditions > 1, n eSNP-eGene pairs (n = 'eqtl.condition.specific') are 
randomly assigned as condition specific.
}
