% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopDesignBatches}
\alias{splatPopDesignBatches}
\title{Set up pooled experimental design}
\usage{
splatPopDesignBatches(params, samples)
}
\arguments{
\item{params}{SplatParams object with simulation parameters.}

\item{samples}{List of samples from vcf.}
}
\value{
Vector with batch assignments for each sample.
}
\description{
Set up pooled experimental design
}
