% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopDesignConditions}
\alias{splatPopDesignConditions}
\title{Set up designed experiments conditions}
\usage{
splatPopDesignConditions(params, samples)
}
\arguments{
\item{params}{SplatParams object with simulation parameters.}

\item{samples}{List of samples from vcf.}
}
\value{
Vector with condition assignments for each sample.
}
\description{
Set up designed experiments conditions
}
