% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopeQTLEffects}
\alias{splatPopeQTLEffects}
\title{Assign eGenes-eSNPs pairs and effect sizes.}
\usage{
splatPopeQTLEffects(params, key, snps)
}
\arguments{
\item{params}{splatPopParams object containing parameters for population
scale simulations. See \code{\link{splatPopParams}} for details.}

\item{key}{Partial splatPop key data.frame.}

\item{snps}{Data.frame of genotype information output from `vcf.parsed`.}
}
\value{
The key updated with assigned eQTL effects.
}
\description{
Randomly pairs N genes (eGene) a SNP (eSNP) within the window size
(eqtl.dist) and assigns each pair an effect size sampled from a gamma
distribution parameterized using the effect sizes from a real eQTL study.
}
