% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopParseGenes}
\alias{splatPopParseGenes}
\title{Generate population key matrix from random or gff provided gene information}
\usage{
splatPopParseGenes(params, gff)
}
\arguments{
\item{params}{splatPopParams object containing parameters for population
scale simulations. See \code{\link{splatPopParams}} for details. 
Default=`newSplatPopParams()`.}

\item{gff}{Either FALSE (if random genes desired) or data.frame object 
of a GFF/GTF file.}
}
\value{
The partial eQTL key dataframe.
}
\description{
Generate population key matrix from random or gff provided gene information
}
