% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopParseVCF}
\alias{splatPopParseVCF}
\title{Format and subset genotype data from a VCF file.}
\usage{
splatPopParseVCF(vcf, params)
}
\arguments{
\item{vcf}{VariantAnnotation object containing genotypes of samples.}

\item{params}{SplatPopParams object containing parameters for population
scale simulations. See \code{\link{SplatPopParams}} for details.}
}
\value{
Genotype data.frame
}
\description{
Extract numeric alleles from vcf object and filter out SNPs missing genotype
data or outside the Minor Allele Frequency range in \code{SplatPopParams}.
}
\keyword{internal}
