% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopSimConditionalEffects}
\alias{splatPopSimConditionalEffects}
\title{Add conditional DE effects to means matrix}
\usage{
splatPopSimConditionalEffects(key, means.pop, conditions)
}
\arguments{
\item{key}{Partial splatPop key data.frame.}

\item{means.pop}{matrix or list of matrices with gene means.}

\item{conditions}{array of condition assignments for each sample}
}
\value{
data.frame of gene mean expression levels WITH conditional DE
effects.
}
\description{
Add conditional DE effects to means matrix
}
