% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-faker.R
\name{splatSetGeneMeans}
\alias{splatSetGeneMeans}
\title{Set gene means}
\usage{
splatSetGeneMeans(sim, params, gene_means)
}
\arguments{
\item{sim}{SingleCellExperiment to add gene means to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated gene means.
}
\description{
Set gene means from a given vector. Also simulates outlier expression factors.
Genes with an outlier factor not equal to 1 are replaced
with the median mean expression multiplied by the outlier factor.
}
