% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-simulate.R
\name{splatSimDE}
\alias{splatSimDE}
\alias{splatSimGroupDE}
\alias{splatSimPathDE}
\title{Simulate group differential expression}
\usage{
splatSimGroupDE(sim, params)

splatSimPathDE(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add differential expression to.}

\item{params}{splatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated differential expression.
}
\description{
Simulate differential expression. Differential expression factors for each
group are produced using \code{\link{getLNormFactors}} and these are added
along with updated means for each group. For paths care is taken to make sure
they are simulated in the correct order.
}
