% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-simulate.R
\name{splatSimLibSizes}
\alias{splatSimLibSizes}
\title{Simulate library sizes}
\usage{
splatSimLibSizes(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add library size to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated library sizes.
}
\description{
Simulate expected library sizes from a log-normal distribution
}
