% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-simulate.R
\name{splatSimulateeQTL}
\alias{splatSimulateeQTL}
\title{Simulate single cell data for eQTL population}
\usage{
splatSimulateeQTL(
  params = newSplatParams(),
  method = "single",
  eqtl = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{params}{SplatParams object with simulation parameters.}

\item{method}{which simulation method to use. Options are "single" which
produces a single population, "groups" which produces distinct groups
(eg. cell types), "paths" which selects cells from continuous
trajectories (eg. differentiation processes).}

\item{eqtl}{Matrix of gene means for eQTL population. Output from 
eQTLSimulate().}

\item{verbose}{logical. Whether to print progress messages.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SingleCellExperiment with simulated data for whole eQTL population.
}
\description{
Simulate count data for an eQTL population from a fictional single-cell 
RNA-seq experiment using the Splat method.
}
