% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-simulate.R
\name{splatSimulatePop}
\alias{splatSimulatePop}
\title{Simulate single cell data for a population}
\usage{
splatSimulatePop(
  params = newSplatParams(),
  method = c("single", "groups", "paths"),
  pop = NULL,
  small = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{params}{SplatParams object with simulation parameters.}

\item{method}{which simulation method to use. Options are "single" which
produces a single cell population for each sample, "groups" which 
produces distinct groups (eg. cell types) for each sample (note, this
creates separate groups from those created in `popSimulate` with only
DE effects), and "paths" which selects cells from continuous
trajectories (eg. differentiation processes).}

\item{pop}{Matrix or list of matrices of gene means for the population. 
Output from `popSimulate()`.}

\item{small}{logical. Whether to return only the counts (Default=FALSE).}

\item{verbose}{logical. Whether to print progress messages.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SingleCellExperiment with simulated data for whole population.
}
\description{
Simulate count data for a population from a fictional single-cell 
RNA-seq experiment using the Splat method.
}
