% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter-package.R
\docType{package}
\name{splatter-package}
\alias{splatter}
\alias{splatter-package}
\title{splatter: Simple Simulation of Single-cell RNA Sequencing Data}
\description{
Splatter is a package for the simulation of single-cell RNA sequencing count data. It provides a simple interface for creating complex simulations that are reproducible and well-documented. Parameters can be estimated from real data and functions are provided for comparing real and simulated datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Oshlack/splatter}
  \item Report bugs at \url{https://github.com/Oshlack/splatter/issues}
}

}
\author{
\strong{Maintainer}: Luke Zappia \email{luke@lazappi.id.au} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID})

Authors:
\itemize{
  \item Belinda Phipson \email{belinda.phipson@petermac.org} (\href{https://orcid.org/0000-0002-1711-7454}{ORCID})
  \item Alicia Oshlack \email{alicia.oshlack@petermac.org} (\href{https://orcid.org/0000-0001-9788-5690}{ORCID})
}

Other contributors:
\itemize{
  \item Christina Azodi \email{cazodi@svi.edu.au} (\href{https://orcid.org/0000-0002-6097-606X}{ORCID}) [contributor]
}

}
\keyword{internal}
