% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter-package.R
\docType{package}
\name{splatter}
\alias{splatter}
\title{splatter.}
\description{
\pkg{splatter} is a package for the well-documented and reproducible
simulation of single-cell RNA-seq count data.
}
\details{
As well as it's own simulation model \pkg{splatter} provides functions for
the estimation of model parameters.
}
\seealso{
Zappia L, Phipson B, Oshlack A. Splatter: Simulation Of Single-Cell RNA
Sequencing Data. bioRxiv. 2017; doi:10.1101/133173
}
