% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-estimate.R
\name{splotchEstBCV}
\alias{splotchEstBCV}
\title{Estimate Splotch BCV parameters}
\usage{
splotchEstBCV(counts, params, verbose)
}
\arguments{
\item{counts}{counts matrix.}

\item{params}{SplotchParams object to store estimated values in.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SplotchParams object with estimated BCV parameters
}
\description{
Estimate Biological Coefficient of Variation (BCV) parameters for the Splotch
simulation
}
\details{
The \code{\link[edgeR]{estimateDisp}} function is used to estimate the common
dispersion across the dataset. An exponential correction is applied based on
fitting an exponential relationship between simulated and estimated values.
If this results in a negative dispersion a simpler linear correction is
applied instead.
}
