% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-estimate.R
\name{splotchEstimate}
\alias{splotchEstimate}
\alias{splotchEstimate.SingleCellExperiment}
\alias{splotchEstimate.matrix}
\title{Estimate Splotch simulation parameters}
\usage{
splotchEstimate(counts, params = newSplotchParams(), verbose = TRUE)

\method{splotchEstimate}{SingleCellExperiment}(counts,
  params = newSplotchParams(), verbose = TRUE)

\method{splotchEstimate}{matrix}(counts, params = newSplotchParams(),
  verbose = TRUE)
}
\arguments{
\item{counts}{either a counts matrix or a SingleCellExperiment object
containing count data to estimate parameters from.}

\item{params}{SplotchParams object to store estimated values in.}

\item{verbose}{logical. Whether to print progress messages.}
}
\value{
SplotchParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the Splotch simulation from a real
dataset. See the individual estimation functions for more details on how this
is done.
}
\examples{
# Load example data
library(scater)
set.seed(1)
sce <- mockSCE()

params <- splotchEstimate(sce)
params
}
\seealso{
\code{\link{splotchEstMean}}, \code{\link{splotchEstLib}}
}
