% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{splotchGenNetwork}
\alias{splotchGenNetwork}
\title{Generate Splotch gene network}
\usage{
splotchGenNetwork(params, verbose)
}
\arguments{
\item{params}{SplotchParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SplotchParams object with gene network
}
\description{
Generate a gene network for the Splotch simulation
}
\details{
Currently a very simple approach is used which needs to be improved. A
network is generated using the \code{\link[igraph]{sample_forestfire}}
function and edge weights are sampled from a standard normal distribution.
}
