% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{splotchSelectRegs}
\alias{splotchSelectRegs}
\title{Select Splotch regulators}
\usage{
splotchSelectRegs(params, verbose)
}
\arguments{
\item{params}{SplotchParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SplotchParams object with gene regulators
}
\description{
Select regulator genes in the gene network for a Splotch simulation
}
\details{
Regulators are randomly selected, weighted according to the difference
between their out degree and in degree. This is an arbitary weighting and
may be improved or replace in the future.
}
