% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{splotchSetup}
\alias{splotchSetup}
\title{Splotch setup}
\usage{
splotchSetup(params = newSplotchParams(), verbose = TRUE, ...)
}
\arguments{
\item{params}{SplotchParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
A complete SplotchParams object
}
\description{
Setup the parameters required for the Splotch simulation
}
\details{
The first stage is a two-step Splotch simulation is to generate some of the
intermediate parameters. The resulting parameters allow multiple simulated
datasets to be generated from the same biological structure (using
\code{\link{splotchSample}}). As with all the other parameters these values
can be manually overwritten if desired.

The setup involves the following steps:
\enumerate{
    \item Generate a gene network (if not already present)
    \item Select regulator genes (if not already present)
    \item Simulate gene means (if not already present)
    \item Simulate cell paths
}

The resulting \code{\link{SplotchParams}} object will have the following
parameters set (if they weren't already).

\itemize{
    \item \code{mean.values}
    \item \code{network.graph}
    \item \code{network.regsSet}
    \item \code{paths.means}
}

See \code{\link{SplotchParams}} for more details about these parameters and
the functions for the individual steps for more details about the process.
}
\examples{
params <- splotchSetup()
}
\seealso{
\code{\link{splotchGenNetwork}}, \code{\link{splotchSelectRegs}},
\code{\link{splotchSimGeneMeans}}, \code{\link{splotchSimPaths}},
\code{\link{SplotchParams}}
}
