% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{splotchSimAmbientCounts}
\alias{splotchSimAmbientCounts}
\title{Simulate Splotch ambient counts}
\usage{
splotchSimAmbientCounts(sim, params, verbose)
}
\arguments{
\item{sim}{SingleCellExperiment containing simulation.}

\item{params}{SplotParams object with simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SingleCellExperiment with ambient counts
}
\description{
Simulate Splotch ambient counts
}
\details{
The overall expression profile to calculated by averaging the cell counts
of the (non-empty) cells. This is then multiplied by the ambient library
sizes to get a mean for each cell. Counts are then sampled from a Poisson
distribution using these means.
}
