% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{splotchSimGeneMeans}
\alias{splotchSimGeneMeans}
\title{Simulate Splotch gene means}
\usage{
splotchSimGeneMeans(params, verbose)
}
\arguments{
\item{params}{SplotchParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}
}
\value{
SplotchParams object with gene means
}
\description{
Simulate Splotch gene means
}
\details{
Gene means are simulated in one of two ways depending on the value of the
\code{mean.method} parameter.

If \code{mean.method} is "fit" (default) then means are sampled from a Gamma
distribution with shape equals \code{mean.shape} and rate equals
\code{mean.rate}. Expression outliers are then added by replacing some
values with the median multiplied by a factor from a log-normal distribution.
This is the same process used for the Splat simulation.

If \code{mean.method} is "density" then means are sampled from the
density object in the \code{mean.density} parameter using a rejection
sampling method. This approach is more flexible but may violate some
statistical assumptions.
}
