% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{splotchSimulate}
\alias{splotchSimulate}
\title{Splotch simulation}
\usage{
splotchSimulate(params = newSplotchParams(), verbose = TRUE, ...)
}
\arguments{
\item{params}{SplotchParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SingleCellExperiment containing simulated counts and intermediate
values
}
\description{
Simulate scRNA-seq count data using the Splotch model
}
\details{
This functions is for simulating data in a single step. It consists of a
call to \code{\link{splotchSetup}} followed by a call to
\code{\link{splotchSample}}. Please see the documentation for those functions
for more details of the individual steps.
}
\examples{
sim <- splotchSimulate()

}
\seealso{
\code{\link{splotchSetup}}, \code{\link{splotchSample}}
}
