% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{summariseDiff}
\alias{summariseDiff}
\title{Summarise diffSCESets}
\usage{
summariseDiff(diff)
}
\arguments{
\item{diff}{Output from \code{\link{diffSCESets}}}
}
\value{
List with MADs, ranks and both combined in long format
}
\description{
Summarise the results of \code{\link{diffSCESets}}. The various
properties are sorted, differences calculated, the Median Absolute Deviation
taken as the summary statistic and the ranks calculated.
}
\examples{
sim1 <- splatSimulate(nGenes = 1000, batchCells = 20)
sim2 <- simpleSimulate(nGenes = 1000, nCells = 20)
difference <- diffSCESets(list(Splat = sim1, Simple = sim2), ref = "Simple")
summary <- summariseDiff(difference)
names(summary)
head(summary$Long)
}
