% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{summariseKS}
\alias{summariseKS}
\title{Summarise KS}
\usage{
summariseKS(data, ref, split.col, stat.cols)
}
\arguments{
\item{data}{The data.frame to summarise}

\item{ref}{The reference data.frame}

\item{split.col}{Name of the column used to split the dataset}

\item{stat.cols}{Names of the columns to summarise. If this vector is named
those names will be used in the output.}
}
\value{
data.frame with the summarised measure, scaled and ranked
}
\description{
Summarise columns of a data.frame compared to a reference using the KS test.
}
