% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{summariseStats}
\alias{summariseStats}
\title{Summarise statistics}
\usage{
summariseStats(data, split.col, stat.cols, measure = c("MAD", "MAE", "RMSE"))
}
\arguments{
\item{data}{The data.frame to summarise}

\item{split.col}{Name of the column used to split the dataset}

\item{stat.cols}{Names of the columns to summarise. If this vector is named
those names will be used in the output.}

\item{measure}{The measure to use for summarisation.}
}
\value{
data.frame with the summarised measure, scaled and ranked
}
\description{
Summarise columns of a data.frame using a single measure.
}
