% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamComparisonModule.R
\name{teamComparisonUI}
\alias{teamComparisonUI}
\title{Module for the team composition comparison page}
\usage{
teamComparisonUI(id)
}
\description{
Module for the team composition comparison page
}
