% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{tidyStatSumm}
\alias{tidyStatSumm}
\title{Tidy summarised statistics}
\usage{
tidyStatSumm(stat.summ, measure = c("MAD", "MAE", "RMSE"))
}
\arguments{
\item{stat.summ}{The summary to convert}

\item{measure}{The name of the summarisation measure}
}
\value{
tidy data.frame with the summarised measure, scaled and ranked
}
\description{
Convert a statistic summary to tidy format and add ranks and scaled values
}
