% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize vector}
\usage{
winsorize(x, q)
}
\arguments{
\item{x}{Numeric vector to winsorize}

\item{q}{Percentile to set from each end}
}
\value{
Winsorized numeric vector
}
\description{
Set outliers in a numeric vector to a specified percentile.
}
