% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCSParams}
\alias{BASiCSParams}
\alias{BASiCSParams-class}
\title{The BASiCSParams class}
\description{
S4 class that holds parameters for the BASiCS simulation.
}
\section{Parameters}{


The BASiCS simulation uses the following parameters:

\describe{
    \item{\code{nGenes}}{The number of genes to simulate.}
    \item{\code{nCells}}{The number of cells to simulate.}
    \item{\code{[seed]}}{Seed to use for generating random numbers.}
}

The parameters not shown in brackets can be estimated from real data using
\code{\link{BASiCSEstimate}}. For details of the BASiCS simulation see
\code{\link{BASiCSSimulate}}.
}

