% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.assign.group.effects}
\alias{eqtl.assign.group.effects}
\title{Designate subset of eQTL associations as group-specific.}
\usage{
eqtl.assign.group.effects(key, groups, eQTLparams, params)
}
\arguments{
\item{key}{Partial eQTL key dataframe.}

\item{groups}{array of group names}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details. Default=`neweQTLParams()`.}

\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details. Default=`newSplatParams()`.}
}
\value{
he key updated with group eQTL and non-eQTL effects.
}
\description{
If groups > 1, n eSNP-eGene pairs (n = 'eqtl.group.specific') are randomly
assigned as group specific.
}
