% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.key.update.quan.norm}
\alias{eqtl.key.update.quan.norm}
\title{Add gene mean and cv after quantile normalization to the eQTL key}
\usage{
eqtl.key.update.quan.norm(key, MeansMatrix)
}
\arguments{
\item{MeansMatrix}{The output from eqtl.quan.norm.sc as a matrix or list of
matrices.}

\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}
}
\value{
Updated key
}
\description{
Add gene mean and cv after quantile normalization to the eQTL key
}
