% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.parse.gff}
\alias{eqtl.parse.gff}
\title{Generate eQTL key matrix using information from the GTF/GFF file.}
\usage{
eqtl.parse.gff(gff)
}
\arguments{
\item{gff}{Dataframe of genes to simulate from a GFF or GTF file.}
}
\value{
The partial eQTL key dataframe.
}
\description{
Generate eQTL key matrix using information from the GTF/GFF file.
}
