% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.quan.norm.sc}
\alias{eqtl.quan.norm.sc}
\title{Quantile normalize expression levels by sample to fit sc parameters}
\usage{
eqtl.quan.norm.sc(params, MeansMatrix)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{MeansMatrix}{The gene expression means for the population}
}
\value{
Matrix of simulated gene means for eQTL population.
}
\description{
Quantile normalize expression levels by sample to fit sc parameters
}
\details{
For each sample, expression value is quantile normalized (qgamma) using the
gamma distribution parameterized from splatEstimate().
}
