% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.sim.eqtl.eff}
\alias{eqtl.sim.eqtl.eff}
\title{Add eQTL effects to means matrix}
\usage{
eqtl.sim.eqtl.eff(id, key, snps, MeansPop)
}
\arguments{
\item{key}{Partial eQTL key dataframe.}

\item{snps}{Dataframe of genotype information output from `vcf.parsed`.}

\item{MeansPop}{Mean gene expression dataframe.}
}
\value{
Dataframe of gene mean expression levels WITH eQTL effects.
}
\description{
Add eQTL effects and non-eQTL group effects to simulated means matrix.
The eQTL effects are incorporated using the following equation:
\deqn{Ygs = (ESg * Mg) + Mgs } 
Where Ygs is the mean for gene *g* and sample *s*, ESg is the effect size
assigned to *g*, Mg is the mean expression assigned to *g*, and Mgs is the
mean expression sampled for *g* for *s*. Non-eQTL group effects are 
incorporated as: 
\deqn{Ygs = Mgs * GEg}
Where GEg is the group effect (i.e. differential expression) assigned to *g*.
To simulate multiple gene mean matrices with different group effects, this
function can be run with `id` designating the group id.
}
\details{
#' @param id The group ID (e.g. "global" or "g1")
}
