% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eqtl.sim.means}
\alias{eqtl.sim.means}
\title{Simulate mean gene expression matrix without eQTL effects}
\usage{
eqtl.sim.means(samples, key)
}
\arguments{
\item{samples}{Vector containing the sample names.}

\item{key}{Partial eQTL key dataframe.}
}
\value{
Dataframe of gene mean expression levels WITHOUT eQTL effects.
}
\description{
Gene mean expression levels are assigned to each gene for each pair randomly
from a normal distribution parameterized using the mean and cv assigned to 
each gene in the key.
}
