% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLMeansMatrix}
\alias{eQTLMeansMatrix}
\title{Project normalized gene expression matrix into mean gene expression matrix}
\usage{
eQTLMeansMatrix(id, key, nGeneMeansPop, last_MeansMatrix)
}
\arguments{
\item{id}{The group ID (e.g. "g1")}

\item{key}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}

\item{nGeneMeansPop}{The normalized gene expression means for the population}

\item{last_MeansMatrix}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}
}
\value{
MeansMatrix: Matrix of simulated gene means for eQTL population.
}
\description{
Project normalized gene expression matrix into mean gene expression matrix
}
\details{
For each gene/sample, the normalized expression value (from rnorm) is 
transformed to the cumulative density function (pnorm) between 0 and 1, this
value is then quantile normalized (qnorm) using the gamma distribution 
parameterized from splatEstimate().
}
