% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLSimulate}
\alias{eQTLSimulate}
\title{eQTL-Simulate}
\usage{
eQTLSimulate(
  params = newSplatParams(),
  eQTLparams = neweQTLParams(),
  gff = ex_gff,
  vcf = ex_snps,
  key = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details.}

\item{gff}{Dataframe containing the genes to include in GFF/GTF format.}

\item{vcf}{Dataframe containing real/simulated genotype data in .vcf format. 
Where each column is a sample and each row is a SNP.}

\item{key}{Path to file with eQTL key information. If FALSE (default), will 
generate new key using params in eQTLparams and genes from gff}

\item{verbose}{logical. Whether to print progress messages.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
A list containing: `means` a dataframe or list of dataframes (if 
n.groups > 1) with the simulated mean gene expression value for each gene 
(row) and each sample (column) and `key` a dataframe with the eSNP-eGene
assignments.
}
\description{
Simulate mean gene counts for a population of samples, such that a defined 
number of associations (cis-eQTL) between markers (eSNPs) and genes 
(eGenes) exist.
}
\details{
Parameters can be set in a variety of ways. If no parameters are 
provided the default parameters are used. Any parameters in \code{params} or
\code{eQTLparams} can be overridden by supplying additional arguments through 
a call to \code{\link{setParams}}. This design allows the user flexibility in
how they supply parameters and allows small adjustments without creating a
new \code{SplatParams} object. See examples for a demonstration of how this
can be used.

eQTLSimulate involves the following steps:
\enumerate{
    \item Read and format genotype data given a .vcf.
    \item Read in genes from GTF/GFF or from a provided key.
    \item If not provided in key, assign expression mean and cv to each gene.
    \item If not provided in key, assign eQTL and effect sizes to n genes. 
    \item Assign gene expression values for each individual for each gene 
    by sampling randomly from a normal distribution parameterized from 
    eqtlEstimate.
    \item Add in eQTL effects. 
    \item Quantile normalize expression levels to match distribution of 
    single cell data from `splatEstimate`.}
}
\examples{
# Load example data
data(ex_gff)
data(ex_snps)
pop.sim <- eQTLSimulate()

}
\seealso{
\code{\link{eqtl.parse.vcf}}, \code{\link{eqtl.parse.snps}},
\code{\link{eqtl.assign.means}}, \code{\link{eqtl.assign.eqtl.effects}}, 
\code{\link{eqtl.assign.group.effects}}, \code{\link{eqtl.sim.means}}
\code{\link{eqtl.sim.eqtl.eff}}, \code{\link{eqtl.quan.norm.sc}}
}
