% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl-simulate.R
\name{eQTLSimulate}
\alias{eQTLSimulate}
\title{eQTL-Simulate}
\usage{
eQTLSimulate(
  params = newSplatParams(),
  eQTLparams = neweQTLParams(),
  gff = ex_gff,
  vcf = ex_snps,
  eqtl.save = TRUE,
  save.name = "default",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{eQTLparams}{eQTLParams object containing parameters for the 
 simulation of the mean expression levels for the population.
See \code{\link{eQTLParams}} for details.}

\item{gff}{Dataframe containing the genes to include in GFF/GTF format.}

\item{vcf}{Dataframe containing real/simulated genotype data in .vcf format. 
Where each column is a sample and each row is a SNP.}

\item{eqtl.save}{logical. Whether to save eQTL key and mean matrix.}

\item{save.name}{String to save eQTL key under. Default: 
eqtl_out/YYMMDD_HHMM_eQTL_key.csv}

\item{verbose}{logical. Whether to print progress messages.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
GeneMeansPop Matrix containing the simulated mean gene expression
value for each gene (row) and each sample in the population (column).
intermediate values.
}
\description{
Simulate mean gene counts for a population of samples, such that a defined 
number of associations (cis-eQTL) between markers (eSNPs) and genes 
(eGenes) exist.
}
\details{
Parameters can be set in a variety of ways. If no parameters are provided
the default parameters are used. Any parameters in \code{params} can be
overridden by supplying additional arguments through a call to
\code{\link{setParams}}. This design allows the user flexibility in
how they supply parameters and allows small adjustments without creating a
new \code{SplatParams} object. See examples for a demonstration of how this
can be used.

The eQTL Gene Mean simulation involves the following steps:
\enumerate{
    \item Load and format gene (GFF/GTF) and SNP (genotype) data.
    \item Select eGenes-eSNPs pairs and assign effect sizes.
    \item Generate normalized gene mean expression matrix for the population. 
    \item Set a gene mean expression value (not normalized) for each gene. 
    \item Generate a gene mean expression matrix for the population.
    \item (optional) Save eQTL key
}
}
\examples{
# Load example data
data(ex_gff)
data(ex_snps)
pop.gMeans <- eQTLSimulate()

}
\seealso{
\code{\link{eQTLgenes}}, \code{\link{eQTLsnps}},
\code{\link{eQTLkey}}, \code{\link{eQTLnormMeansMatrix}},
\code{\link{eQTLGeneMeans}}, \code{\link{eQTLMeansMatrix}}
}
