% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-estimate.R
\name{estimateSimpleParams}
\alias{estimateSimpleParams}
\title{Estimate simple simulation parameters}
\usage{
estimateSimpleParams(data, params = newSimpleParams())
}
\arguments{
\item{params}{SimpleParams object to store estimated values in.}

\item{x}{either a counts matrix or an SCESet object containing count data to
estimate parameters from.}
}
\value{
SimpleParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the simple simulation from a real dataset.
}
\details{
The \code{nGenes} and \code{nCells} parameters are taken from the size of the
input data. The mean parameters are estimated by fitting a gamma distribution
to the library size normalised mean expression level using
\code{\link[fitdistrplus]{fitdist}}. See \code{\link{SimpleParams}} for more
details on the parameters.
}
\examples{
data("sc_example_counts")
params <- estimateSimpleParams(sc_example_counts)
params
}

