% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estMeanParams}
\alias{estMeanParams}
\title{Estimate mean parameters}
\usage{
estMeanParams(norm.counts, params)
}
\arguments{
\item{norm.counts}{library size normalised counts matrix.}

\item{params}{splatParams object to store estimated values in.}
}
\value{
splatParams object with estimated values.
}
\description{
Estimate rate and shape parameters for the gamma distribution used to
simulate gene expression means using the 'moment matching estimation' method
of \code{\link[fitdistrplus]{fitdist}}.
}
\examples{
\dontrun{
data("sc_example_counts")
norm_ex_counts <- t(t(sc_example_counts) / colSums(sc_example_counts) *
                  median(colSums(sc_example_counts)))
params <- splatParams()
params <- estMeanParams(norm_ex_counts, params)
params
}
}

