% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BASiCSParams-methods.R,
%   R/LunParams-methods.R, R/SplatParams-methods.R, R/SplatPopParams-methods.R
\name{expandParams}
\alias{expandParams}
\alias{expandParams,BASiCSParams-method}
\alias{expandParams,LunParams-method}
\alias{expandParams,SplatParams-method}
\alias{expandParams,SplatPopParams-method}
\title{Expand parameters}
\usage{
expandParams(object, ...)

\S4method{expandParams}{BASiCSParams}(object)

\S4method{expandParams}{LunParams}(object)

\S4method{expandParams}{SplatParams}(object)

\S4method{expandParams}{SplatPopParams}(object)
}
\arguments{
\item{object}{object to expand.}

\item{...}{additional arguments.}
}
\value{
Expanded object.
}
\description{
Expand the parameters that can be vectors so that they are the same length as
the number of groups.
}
