% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{expandParams}
\alias{expandParams}
\title{Expath parameters}
\usage{
expandParams(params)
}
\arguments{
\item{params}{splatParams object to expand.}
}
\value{
expanded splatParams object.
}
\description{
Expand the parameters that can be vecotor so that they are the same length as
the number of groups.
}
\examples{
\dontrun{
params <- defaultParams()
params <- setParams(params, groupCells = c(10, 10))
params
params <- expandPathParams(params)
params
}
}

