% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{expandParams}
\alias{expandParams}
\title{Expand parameters}
\usage{
expandParams(object, ...)
}
\arguments{
\item{object}{object to expand.}

\item{...}{additional arguments.}
}
\value{
Expanded object.
}
\description{
Expand the parameters that can be vectors so that they are the same length as
the number of groups.
}
\examples{
params <- newSplatParams()
params <- setParams(params, groupCells = c(10, 10))
params
params <- expandParams(params)
params
}

