% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BASiCSParams-methods.R,
%   R/LunParams-methods.R, R/Params-methods.R, R/SplatParams-methods.R,
%   R/SplatPopParams-methods.R, R/params-functions.R
\name{expandParams}
\alias{expandParams}
\alias{expandParams,BASiCSParams-method}
\alias{expandParams,LunParams-method}
\alias{expandParams,Params-method}
\alias{expandParams,SplatParams-method}
\alias{expandParams,SplatPopParams-method}
\alias{paramsExpander}
\title{Expand parameters}
\usage{
expandParams(object, ...)

\S4method{expandParams}{BASiCSParams}(object)

\S4method{expandParams}{LunParams}(object)

\S4method{expandParams}{Params}(object, vectors, n)

\S4method{expandParams}{SplatParams}(object)

\S4method{expandParams}{SplatPopParams}(object)

paramsExpander(object, vectors, n)
}
\arguments{
\item{object}{object to expand.}

\item{...}{additional arguments.}

\item{n}{number of times to repeat each parameter}

\item{vector}{names of vector parameters to expand}
}
\value{
Expanded object.
}
\description{
Expand the parameters that can be vectors so that they are the same length as
the number of groups. Work is done by \code{\link{paramsExpander}} called
from each method. Expansions are stored using
\code{\link{setParamsUnchecked}}.
}
