% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Lun2Params}
\alias{Lun2Params}
\alias{Lun2Params-class}
\title{The Lun2Params class}
\description{
S4 class that holds parameters for the Lun simulation.
}
\section{Parameters}{


The Lun2 simulation uses the following parameters:

\describe{
  \item{\code{nGenes}}{The number of genes to simulate.}
  \item{\code{nCells}}{The number of cells to simulate.}
  \item{\code{[seed]}}{Seed to use for generating random numbers.}
}

The parameters not shown in brackets can be estimated from real data using
\code{\link{lun2Estimate}}. For details of the Lun2 simulation see
\code{\link{lun2Simulate}}.
}

