% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{makeDiffPanel}
\alias{makeDiffPanel}
\title{Make difference panel}
\usage{
makeDiffPanel(comp, labels = c("A", "B", "C", "D", "E", "F", "G", "H", "I",
  "J", "K", "L", ""))
}
\arguments{
\item{comp}{list returned by \code{\link{diffSCESets}}.}
}
\value{
Combined panel plot
}
\description{
Combine the plots from \code{diffSCESets} into a single panel.
}
\examples{
sim1 <- splatSimulate(nGenes = 1000, groupCells = 20)
sim2 <- simpleSimulate(nGenes = 1000, nCells = 20)
difference <- diffSCESets(list(Splat = sim1, Simple = sim2), ref = "Simple")
panel <- makeCompPanel(difference)

}
