% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{makeOverallPanel}
\alias{makeOverallPanel}
\title{Make overall panel}
\usage{
makeOverallPanel(comp, diff, title = "Overall comparison",
  row.labels = c("Means", "Variance", "Mean-variance relationship",
  "Library size", "Zeros per cell", "Zeros per gene",
  "Mean-zeros relationship"))
}
\arguments{
\item{comp}{list returned by \code{\link{compareSCESets}}.}

\item{diff}{list returned by \code{\link{diffSCESets}}.}

\item{title}{title for the panel.}

\item{row.labels}{vector of labels for each of the seven rows.}
}
\value{
Combined panel plot
}
\description{
Combine the plots from \code{compSCESets} and \code{diffSCESets} into a
single panel.
}
\examples{
\dontrun{
sim1 <- splatSimulate(nGenes = 1000, groupCells = 20)
sim2 <- simpleSimulate(nGenes = 1000, nCells = 20)
comparison <- compSCESets(list(Splat = sim1, Simple = sim2))
difference <- diffSCESets(list(Splat = sim1, Simple = sim2), ref = "Simple")
panel <- makeOverallPanel(comparison, difference)
}

}
