% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mockGFF}
\alias{mockGFF}
\title{Generate mock gff}
\usage{
mockGFF(n.genes = 50, chromosome = 1, chr.length = 2e+06, seed = NULL)
}
\arguments{
\item{n.genes}{Number of genes in mock gff file}

\item{chromosome}{Chromosome name}

\item{chr.length}{Length of mock chromosome}

\item{seed}{Optional: seed for random seed}
}
\value{
data.frame containing mock gff data.
}
\description{
Quick function to generate a mock gff.
}
\examples{
gff <- mockGFF()

}
