% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mockGFF}
\alias{mockGFF}
\title{Generate mock gff}
\usage{
mockGFF(n.genes = 500, chromosome = 22)
}
\arguments{
\item{n.genes}{Number of genes in mock gff file}

\item{chromosome}{Chromosome name}
}
\value{
data.frame containing mock gff data.
}
\description{
Quick function to generate a mock gff.
}
\examples{
gff <- mockGFF()

}
