% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-data.R
\name{mockVCF}
\alias{mockVCF}
\title{Generate mock vcf}
\usage{
mockVCF(n.snps = 10000, n.samples = 10, chromosome = 22)
}
\arguments{
\item{n.snps}{Number of SNPs in mock vcf file.}

\item{n.samples}{Number of samples in mock bulk data.}

\item{chromosome}{Chromosome name}
}
\value{
data.frame containing mock gff data.
}
\description{
Quick function to generate mock vcf file. Note this data has unrealistic
population structure.
}
\examples{
vcf <- mockVCF()

}
