% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/LunParams-methods.R, R/SimpleParams-methods.R, R/SplatParams-methods.R
\name{newParams}
\alias{newLunParams}
\alias{newParams}
\alias{newSimpleParams}
\alias{newSplatParams}
\title{New Params}
\usage{
newLunParams(...)

newSimpleParams(...)

newSplatParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}
}
\value{
New Params object.
}
\description{
Create a new Params object. Functions exist for each of the different
Params subtypes.
}
\examples{
params <- newSimpleParams()
params <- newSimpleParams(nGenes = 200, ncells = 10)

}

