% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Lun2Params-methods.R,
%   R/LunParams-methods.R, R/MFAParams-methods.R, R/SCDDParams-methods.R,
%   R/SimpleParams-methods.R, R/SplatParams-methods.R
\name{newParams}
\alias{newParams}
\alias{newLun2Params}
\alias{newLunParams}
\alias{newMFAParams}
\alias{newSCDDParams}
\alias{newSimpleParams}
\alias{newSplatParams}
\title{New Params}
\usage{
newLun2Params(...)

newLunParams(...)

newMFAParams(...)

newSCDDParams(...)

newSimpleParams(...)

newSplatParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}
}
\value{
New Params object.
}
\description{
Create a new Params object. Functions exist for each of the different
Params subtypes.
}
\examples{
params <- newSimpleParams()
params <- newSimpleParams(nGenes = 200, nCells = 10)

}
