% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BASiCSParams-methods.R,
%   R/KersplatParams-methods.R, R/Lun2Params-methods.R, R/LunParams-methods.R,
%   R/MFAParams-methods.R, R/PhenoParams-methods.R, R/SCDDParams-methods.R,
%   R/SimpleParams-methods.R, R/SparseDCParams-methods.R,
%   R/SplatParams-methods.R, R/ZINBParams-methods.R, R/popParams-methods.R
\name{newParams}
\alias{newParams}
\alias{newBASiCSParams}
\alias{newKersplatParams}
\alias{newLun2Params}
\alias{newLunParams}
\alias{newMFAParams}
\alias{newPhenoParams}
\alias{newSCDDParams}
\alias{newSimpleParams}
\alias{newSparseDCParams}
\alias{newSplatParams}
\alias{newZINBParams}
\alias{newPopParams}
\title{New Params}
\usage{
newBASiCSParams(...)

newKersplatParams(...)

newLun2Params(...)

newLunParams(...)

newMFAParams(...)

newPhenoParams(...)

newSCDDParams(...)

newSimpleParams(...)

newSparseDCParams(...)

newSplatParams(...)

newZINBParams(...)

newPopParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}
}
\value{
New Params object.
}
\description{
Create a new Params object. Functions exist for each of the different
Params subtypes.
}
\examples{
params <- newSimpleParams()
params <- newSimpleParams(nGenes = 200, nCells = 10)

}
