% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Params}
\alias{Params}
\alias{Params-class}
\title{The Params virtual class}
\description{
Virtual S4 class that all other Params classes inherit from.
}
\section{Parameters}{


The Params class defines the following parameters:

\describe{
    \item{\code{nGenes}}{The number of genes to simulate.}
    \item{\code{nCells}}{The number of cells to simulate.}
    \item{\code{[seed]}}{Seed to use for generating random numbers.}
}

The parameters not shown in brackets can be estimated from real data.
}

