% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDD-estimate.R
\name{scDDEstimate}
\alias{scDDEstimate}
\alias{scDDEstimate.SCESet}
\alias{scDDEstimate.matrix}
\title{Estimate scDD simulation parameters}
\usage{
scDDEstimate(counts, conditions, params = newSCDDParams())

\method{scDDEstimate}{SCESet}(counts, conditions, params = newSCDDParams())

\method{scDDEstimate}{matrix}(counts, conditions, params = newSCDDParams())
}
\arguments{
\item{counts}{either a counts matrix or an SCESet object containing count
data to estimate parameters from.}

\item{conditions}{Vector giving the condition that each cell belongs to.
Conditions can be 1 or 2.}

\item{params}{SCDDParams object to store estimated values in.}
}
\value{
SCDDParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the scDD simulation from a real dataset.
}
\details{
This function is just a wrapper around \code{\link[scDD]{preprocess}} that
takes the output and converts it to a SCDDParams object. See
\code{\link[scDD]{preprocess}} for details.
}
\examples{
data("sc_example_counts")
conditions <- sample(1:2, ncol(sc_example_counts), replace = TRUE)
params <- scDDEstimate(sc_example_counts, conditions)
params
}

