% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDD-estimate.R
\name{scDDEstimate}
\alias{scDDEstimate}
\alias{scDDEstimate.SCESet}
\alias{scDDEstimate.matrix}
\title{Estimate scDD simulation parameters}
\usage{
scDDEstimate(counts, conditions, params = newSCDDParams(),
  BPPARAM = SerialParam())

\method{scDDEstimate}{SCESet}(counts, conditions, params = newSCDDParams(),
  BPPARAM = SerialParam())

\method{scDDEstimate}{matrix}(counts, conditions, params = newSCDDParams(),
  BPPARAM = SerialParam())
}
\arguments{
\item{counts}{either a counts matrix or an SCESet object containing count
data to estimate parameters from.}

\item{conditions}{Vector giving the condition that each cell belongs to.
Conditions can be 1 or 2.}

\item{params}{SCDDParams object to store estimated values in.}

\item{BPPARAM}{A \code{\link[BiocParallel]{BiocParallelParam}} instance
giving the parallel back-end to be used. Default is
\code{\link[BiocParallel]{SerialParam}} which uses a single core.}
}
\value{
SCDDParams object containing the estimated parameters.
}
\description{
Estimate simulation parameters for the scDD simulation from a real dataset.
}
\details{
This function applies \code{\link[scDD]{preprocess}} to the counts then uses
\code{\link[scDD]{scDD}} to estimate the numbers of each gene type to
simulate. The output is then converted to a SCDDParams object. See
\code{\link[scDD]{preprocess}} and \code{\link[scDD]{scDD}} for details.
}
\examples{
data("sc_example_counts")
conditions <- sample(1:2, ncol(sc_example_counts), replace = TRUE)
params <- scDDEstimate(sc_example_counts, conditions)
params
}
