% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDD-simulate.R
\name{scDDSimulate}
\alias{scDDSimulate}
\title{scDD simulation}
\usage{
scDDSimulate(params = newSCDDParams(), plots = FALSE, plot.file = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{params}{SCDDParams object containing simulation parameters.}

\item{plots}{logical. whether to generate scDD fold change and validation
plots.}

\item{plot.file}{File path to save plots as PDF.}

\item{verbose}{logical. Whether to print progress messages}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SCESet containing simulated counts
}
\description{
Simulate counts using the scDD method.
}
\details{
This function is just a wrapper around \code{\link[scDD]{simulateSet}} that
takes a \code{\link{SCDDParams}}, runs the simulation then converts the
output to an \code{\link[scater]{SCESet}} object. See
\code{\link[scDD]{simulateSet}} for more details of how the simulation works.
}
\examples{
\dontrun{
sim <- scDDSimulate()
}
}
\references{
Korthauer KD, Chu L-F, Newton MA, Li Y, Thomson J, Stewart R, et al. scDD: A
statistical approach for identifying differential distributions in
single-cell RNA-seq experiments. bioRxiv (2015).

Paper: \url{dx.doi.org/10.1101/035501}

Code: \url{https://github.com/kdkorthauer/scDD}
}

