% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Params-methods.R, R/SplatParams-methods.R
\docType{methods}
\name{setParam}
\alias{setParam}
\alias{setParam,Params-method}
\alias{setParam,SplatParams-method}
\title{Set a parameter}
\usage{
setParam(object, name, value, checkValid)

\S4method{setParam}{Params}(object, name, value, checkValid = TRUE)

\S4method{setParam}{SplatParams}(object, name, value, checkValid)
}
\arguments{
\item{object}{object to set parameter in.}

\item{name}{name of the parameter to set.}

\item{value}{value to set the paramter to.}

\item{checkValid}{logical. Check object is valid after setting.}
}
\value{
Object with new parameter value.
}
\description{
Function for setting parameter values.
}

