% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-functions.R
\name{showPP}
\alias{showPP}
\title{Show pretty print}
\usage{
showPP(params, pp)
}
\arguments{
\item{params}{object to show.}

\item{pp}{list specifying how the object should be displayed.}
}
\value{
Print params object to console
}
\description{
Function used for pretty printing params object.
}
