% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-functions.R
\name{showValues}
\alias{showValues}
\title{Show values}
\usage{
showValues(values, not.default)
}
\arguments{
\item{values}{list of values to show.}

\item{not.default}{logical vector giving which have changed from the default.}
}
\value{
Print values
}
\description{
Function used for pretty printing scalar or vector parameters.
}
