% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-eqtl.R
\name{splateQTLMeansMatrix}
\alias{splateQTLMeansMatrix}
\title{Project normalized gene expression matrix into mean gene expression matrix}
\usage{
splateQTLMeansMatrix(params, pairs, nMeans)
}
\arguments{
\item{params}{SplatParams object containing parameters for the simulation.
See \code{\link{SplatParams}} for details.}

\item{pairs}{A dataframe eSNPs-eGenes pair assignments and their effect sizes}

\item{nMeans}{The normalized gene expression means for the population}
}
\value{
MeansMatrix: Matrix of simulated gene means for eQTL population.
}
\description{
Project normalized gene expression matrix into mean gene expression matrix
}
\details{
For each gene/sample, the normalized expression value (from rnorm) is 
transformed to the cumulative density function (pnorm) between 0 and 1, this
value is then quantile normalized (qgamma) using the gamma distribution 
parameterized from splatEstimate().
}
