% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-estimate.R
\name{splatEstLib}
\alias{splatEstLib}
\title{Estimate Splat library size parameters}
\usage{
splatEstLib(counts, params)
}
\arguments{
\item{counts}{counts matrix to estimate parameters from.}

\item{params}{splatParams object to store estimated values in.}
}
\value{
SplatParams object with estimated values.
}
\description{
The Shapiro-Wilks test is used to determine if the library sizes are
normally distributed. If so a normal distribution is fitted to the library
sizes, if not (most cases) a log-normal distribution is fitted and the
estimated parameters are added to the params object. See
\code{\link[fitdistrplus]{fitdist}} for details on the fitting.
}
