% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{splatParams}
\alias{splatParams}
\title{Create splatParams object}
\usage{
splatParams(...)
}
\arguments{
\item{...}{parameters to set in the new params object, passed to
\code{\link{setParams}}.}
}
\value{
List based S3 splatParams object
}
\description{
S3 class for holding Splatter simulation parameters.
}
\details{
The splatParams object is a list based S3 object for holding simulation
parameters. It has the following sections and values:

\itemize{
  \item nGenes - Number of genes to simulate.
  \item nCells - Number of cells to simulate.
  \item [groupCells] - Vector giving the number of cells in each simulation
        group/path.
  \item mean (mean parameters)
    \itemize{
      \item rate - Rate parameter for the mean gamma simulation.
      \item shape - Shape parameter for the mean gamma simulation.
    }
  \item lib (library size parameters)
    \itemize{
      \item loc - Location (meanlog) parameter for the library size
            log-normal distribution.
      \item scale - Scale (sdlog) parameter for the library size log-normal
            distribution.
    }
  \item out (expression outlier parameters)
    \itemize{
      \item prob - Probability that a gene is an expression outlier.
      \item loProb - Probability that an expression outlier gene is lowly
            expressed.
      \item facLoc - Location (meanlog) parameter for the expression outlier
            factor log-normal distribution.
      \item facScale - Scale (sdlog) parameter for the expression outlier
            factor log-normal distribution.
    }
  \item de (differential expression parameters)
    \itemize{
      \item [prob] - Probability that a gene is differentially expressed
            between groups or paths.
      \item [downProb] - Probability that differentially expressed gene is
            down-regulated.
      \item [facLoc] - Location (meanlog) parameter for the differential
            expression factor log-normal distribution.
      \item [facScale] - Scale (sdlog) parameter for the differential
            expression factor log-normal distribution.
    }
  \item bcv (Biological Coefficient of Variation parameters)
    \itemize{
      \item common - Underlying common dispersion across all genes.
      \item DF - Degrees of Freedom for the BCV inverse chi-squared
            distribution.
    }
  \item dropout (dropout parameters)
    \itemize{
      \item present - Logical. Whether to simulate dropout.
      \item mid - Midpoint parameter for the dropout logistic function.
      \item shape - Shape parameter for the dropout logistic function.
    }
  \item path (differentiation path parameters)
    \itemize{
      \item [from] - Vector giving the originating point of each path. This
            allows path structure such as a cell type which differentiates
            into an intermediate cell type that then differentiates into two
            mature cell types. A path structure of this form would have a
            "from" parameter of c(0, 1, 1) (where 0 is the origin). If no
            vector is given all paths will start at the origin.
      \item [length] - Vector giving the number of steps to simulate along
            each path. If a single value is given it will be applied to all
            paths.
      \item [skew] - Vector giving the skew of each path. Values closer to 1
            will give more cells towards the starting population, values
            closer to 0 will give more cells towards the final population.
            If a single value is given it will be applied to all paths.
      \item [nonlinearProb] - Probability that a gene follows a non-linear
            path along the differentiation path. This allows more complex
            gene patterns such as a gene being equally expressed at the
            beginning an end of a path but lowly expressed in the middle.
      \item [sigmaFac] - Sigma factor for non-linear gene paths. A higher
            value will result in more extreme non-linear variations along a
            path.
  }
}

Those shown in brackets cannot currently be easily estimated from a real
dataset by Splatter. This is also shown when a splatParams object is printed
with parameters labelled as either (estimatable) or [not estimatable].
}
\examples{
params <- splatParams()
params
}

