% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopAssignMeans}
\alias{splatPopAssignMeans}
\title{Sample expression mean and variance for each gene}
\usage{
splatPopAssignMeans(params, key)
}
\arguments{
\item{params}{splatPopParams object containing parameters for population
scale simulations. See \code{\link{splatPopParams}} for details. 
Default=`newSplatPopParams()`.}

\item{key}{Partial eQTL key dataframe.}
}
\value{
The key updated with assigned means and variances.
}
\description{
A mean and coefficient of variation is assigned to each gene by sampling from
gamma distributions parameterized from real data in `splatPopEstimate`.
The cv gamma distributions are binned by gene mean because the distribution
of variance in real data is not independent from the mean.
}
