% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-estimate.R
\name{splatPopEstimateMeanCV}
\alias{splatPopEstimateMeanCV}
\title{Estimate gene mean and gene mean variance parameters}
\usage{
splatPopEstimateMeanCV(params, means)
}
\arguments{
\item{params}{SplatPopParams object containing parameters for the
simulation of the mean expression levels for the population.
See \code{\link{SplatPopParams}} for details.}

\item{means}{Dataframe of real gene means across a population, where
each row is a gene and each column is an individual in the population.}
}
\value{
params object with estimated values.
}
\description{
The Shapiro-Wilks test is used to determine if the library sizes are
normally distributed. If so a normal distribution is fitted to the library
sizes, if not (most cases) a log-normal distribution is fitted and the
estimated parameters are added to the params object. See
\code{\link[fitdistrplus]{fitdist}} for details on the fitting.
}
\details{
Parameters for the mean gamma distribution are estimated by fitting the mean
(across the population) expression of genes that meet the criteria (<50% of
samples have exp <0.1) and parameters for the cv gamma distribution are
estimated for each bin of mean expression using the cv of expression across
the population for genes in that bin. Both are fit using
\code{\link[fitdistrplus]{fitdist}}. The "Nelder-Mead" method is used to fit
the mean gamma distribution and the maximum goodness-of-fit estimation
method is used to minimise the Cramer-von Mises distance for the CV
distribution.
}
