% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopGroupEffects}
\alias{splatPopGroupEffects}
\title{Assign group-specific eQTL and DEGs.}
\usage{
splatPopGroupEffects(params, key, groups)
}
\arguments{
\item{params}{SplatPopParams object containing parameters for population
scale simulations. See \code{\link{SplatPopParams}} for details.}

\item{key}{Partial splatPop key data.frame.}

\item{groups}{array of group names}
}
\value{
The key updated with group eQTL and non-eQTL effects.
}
\description{
If groups > 1, n eSNP-eGene pairs (n = 'eqtl.group.specific') are randomly
assigned as group specific.
}
