% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{splatPopParams}
\alias{splatPopParams}
\alias{splatPopParams-class}
\title{The splatPopParams class}
\description{
S4 class that holds parameters for the splatPop simulation.
}
\section{Parameters}{


In addition to the \code{\link{SplatParams}} parameters, splatPop simulation
 requires the following parameters:

\describe{
    \item{\code{[random.genes]}}{Logical specifying if splatPopSimulateMeans
    should simulate random genes (TRUE) or if it should look for genes in
    the *-gff* or *-key* flag.}
    \item{\code{[group.prop]}}{Proportion of single cells to simulate in each
    group.}
    \item{\code{[eqtl.n]}}{The number (>1) or percent (<=1) of genes to 
    assign eQTL effects.}
    \item{\code{[eqtl.dist]}}{Maximum distance between eSNP and eGene}
    \item{\code{[eqtl.maf.min]}}{Minimum Minor Allele Frequency of eSNPs.}
    \item{\code{[eqtl.maf.max]}}{Maximum Minor Allele Frequency of eSNPs.}
    \item{\code{[eqtl.group.specific]}}{Percent of eQTL effects to simulate 
    as group specific.}
    \item{\emph{eQTL Effect size distribution parameters. Defaults estimated
    from GTEx eQTL mapping results, see vignette for more information.}}{
        \describe{
            \item{\code{eqtl.ES.shape}}{Shape parameter for the effect size 
            gamma distribution.}
            \item{\code{eqtl.ES.rate}}{Rate parameter for the effect size 
            gamma distribution.}
        }
    }
    \item{\emph{Bulk Mean Expression distribution parameters. Defaults 
    estimated from GTEx data, see vignette for more information.}}{
        \describe{
            \item{\code{pop.mean.shape}}{Shape parameter for the mean (i.e. 
            bulk) expression gamma distribution}
            \item{\code{pop.mean.rate}}{Rate parameter for the mean (i.e. 
            bulk) expression gamma distribution}
        }
    }
    \item{\emph{Bulk Expression Coefficient of Variation distribution
    parameters binned. Defaults estimated from GTEx data, see vignette for 
    more information.}}{
        \describe{
            \item{\code{pop.cv.param}}{Dataframe containing gene
            mean bin range, and the CV shape, and CV rate parameters for
            each of those bins.}
        }
    }
}
The parameters not shown in brackets can be estimated from real data using
\code{\link{splatPopEstimate}}. For details of the eQTL simulation
see \code{\link{splatPopSimulate}}.
}

