% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopParseGenes}
\alias{splatPopParseGenes}
\title{Generate population key matrix from random or gff provided gene information}
\usage{
splatPopParseGenes(params, gff)
}
\arguments{
\item{params}{SplatPopParams object containing parameters for population
scale simulations. See \code{\link{SplatPopParams}} for details.}

\item{gff}{Either NULL or a data.frame object containing a GFF/GTF file.}
}
\value{
The Partial splatPop key data.frame.
}
\description{
Generate population key matrix from random or gff provided gene information
}
