% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopQuantNormKey}
\alias{splatPopQuantNormKey}
\title{Add quantile normalized gene mean and cv info the eQTL key.}
\usage{
splatPopQuantNormKey(key, MeansMatrix)
}
\arguments{
\item{key}{Partial splatPop key data.frame.}

\item{MeansMatrix}{The output from `splatPopQuantNorm` as a matrix or list of
matrices.}
}
\value{
Final eQTL key.
}
\description{
Add quantile normalized gene mean and cv info the eQTL key.
}
