% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopQuantNormKey}
\alias{splatPopQuantNormKey}
\title{Add quantile normalized gene mean and cv info the eQTL key.}
\usage{
splatPopQuantNormKey(key, means)
}
\arguments{
\item{key}{Partial splatPop key data.frame.}

\item{means}{matrix or list of matrices containing means from
`splatPopQuantNorm`}
}
\value{
Final eQTL key.
}
\description{
Add quantile normalized gene mean and cv info the eQTL key.
}
