% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopSimulate}
\alias{splatPopSimulate}
\title{splatPop simulation}
\usage{
splatPopSimulate(
  vcf = mock_vcf(),
  params = newSplatPopParams(nGenes = 1000),
  method = c("single", "groups", "paths"),
  gff = FALSE,
  key = "new",
  counts_only = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{vcf}{vcfR object containing genotypes of samples to simulate.}

\item{params}{splatPopParams object containing parameters for population
scale simulations. See \code{\link{splatPopParams}} for details. 
Default=`newSplatPopParams()`.}

\item{method}{which simulation method to use. Options are "single" which
produces a single population, "groups" which produces distinct groups
(eg. cell types), "paths" which selects cells from continuous
trajectories (eg. differentiation processes).}

\item{gff}{Either FALSE (if random.genes=TRUE) or a data.frame object 
containing a GFF/GTF file.}

\item{key}{Either FALSE or a data.frame object containing a full or partial
splatPop key.}

\item{counts_only}{logical. Whether to save only counts in sce object.}

\item{verbose}{logical. Whether to print progress messages.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
A list containing: `means` a dataframe (or list of dataframes if
n.groups > 1) with the simulated mean gene expression value for each gene
(row) and each sample (column), `key` a dataframe with population information
including eQTL and group effects, and `sce` a SingleCellExperiment object
containing simulated counts and intermediate values.
}
\description{
Simulate scRNA-seq count data using the splat model for a population of
individuals with correlation structure.
}
\details{
This functions is for simulating data in a single step. It consists of a
call to \code{\link{splatPopSimulateMeans}}, which simulates a mean 
expression level per gene per sample, followed by a call to 
\code{\link{splatPopSimulateSC}}, which uses the splat model to simulate 
single-cell counts per individual. Please see the documentation for those 
functions for more details.
}
\examples{
if (requireNamespace("vcfR", quietly = TRUE)) {
    library(vcfR)
    vcf <- mock_vcf(n_samples=6)
    sim <- splatPopSimulate(vcf=vcf)
}
sim <- splatPopSimulate()

}
\seealso{
\code{\link{splatPopSimulateMeans}}, \code{\link{splatPopSimulateSC}}
}
