% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatPop-simulate.R
\name{splatPopSimulateSC}
\alias{splatPopSimulateSC}
\title{splatPopSimulateSC}
\usage{
splatPopSimulateSC(
  sim_means,
  params,
  method = c("single", "groups", "paths"),
  counts_only = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sim_means}{Matrix or list of matrices of gene means for the population. 
Output from `splatPopSimulateMeans()`.}

\item{params}{splatPopParams object containing parameters for population
scale simulations. See \code{\link{splatPopParams}} for details. 
Default=`newSplatPopParams()`.}

\item{method}{which simulation method to use. Options are "single" which
produces a single cell population for each sample, "groups" which 
produces distinct groups (eg. cell types) for each sample (note, this
creates separate groups from those created in `popSimulate` with only
DE effects), and "paths" which selects cells from continuous
trajectories (eg. differentiation processes).}

\item{counts_only}{logical. Whether to return only the counts (Default=FALSE).}

\item{verbose}{logical. Whether to print progress messages.}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SingleCellExperiment with simulated data for whole population.
}
\description{
Simulate count data for a population from a fictional single-cell 
RNA-seq experiment using the Splat method.
}
