% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SplotchParams}
\alias{SplotchParams}
\alias{SplotchParams-class}
\title{The SplotchParams class}
\description{
S4 class that holds parameters for the Splotch simulation.
}
\section{Parameters}{


The Splotch simulation uses the following parameters:

\describe{
    \item{\code{nGenes}}{The number of genes to simulate.}
    \item{\code{nCells}}{The number of cells to simulate.}
    \item{\code{[seed]}}{Seed to use for generating random numbers.}
    \item{\emph{Mean parameters}}{
        \describe{
            \item{\code{mean.shape}}{Shape parameter for the mean gamma
            distribution.}
            \item{\code{mean.rate}}{Rate parameter for the mean gamma
            distribution.}
            \item{\code{mean.values}}{Vector of means for each gene.}
        }
    }
    \item{\emph{Network parameters}}{
        \describe{
            \item{\code{[network.graph]}}{Graph containing the gene network.}
            \item{\code{[network.nRegs]}}{Number of regulators in the
            network.}
        }
    }
}

The parameters not shown in brackets can be estimated from real data using
\code{\link{splotchEstimate}}. For details of the Splotch simulation
see \code{\link{splotchSimulate}}.
}

