% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splotch-simulate.R
\name{splotchSimulate}
\alias{splotchSimulate}
\title{Splotch simulation}
\usage{
splotchSimulate(params = newSplotchParams(), verbose = TRUE, ...)
}
\arguments{
\item{params}{SplotchParams object containing simulation parameters.}

\item{verbose}{logical. Whether to print progress messages}

\item{...}{any additional parameter settings to override what is provided in
\code{params}.}
}
\value{
SingleCellExperiment containing simulated counts
}
\description{
Simulate counts from...
}
\details{
Details...
}
\examples{
sim <- splotchSimulate()

}
