% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ZINBParams}
\alias{ZINBParams}
\alias{ZINBParams-class}
\title{The ZINBParams class}
\description{
S4 class that holds parameters for the ZINB-WaVE simulation.
}
\section{Parameters}{


The ZINB-WaVE simulation uses the following parameters:

\describe{
    \item{\code{nGenes}}{The number of genes to simulate.}
    \item{\code{nCells}}{The number of cells to simulate.}
    \item{\code{[seed]}}{Seed to use for generating random numbers.}
    \item{\code{model}}{Object describing a ZINB model.}
}

The majority of the parameters for this simulation are stored in a
\code{\link[zinbwave]{ZinbModel}} object. Please refer to the documentation
for this class and its constructor(\code{\link[zinbwave]{zinbModel}}) for
details about all the parameters.

The parameters not shown in brackets can be estimated from real data using
\code{\link{zinbEstimate}}. For details of the ZINB-WaVE simulation
see \code{\link{zinbSimulate}}.
}

